% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_media.R
\name{ala_media}
\alias{ala_media}
\title{Image and sounds}
\usage{
ala_media(taxa, filters, locations, columns, download_dir)
}
\arguments{
\item{taxa}{\code{data.frame}: generated by a call to \code{\link{select_taxa}()}. This
argument also accepts a vector of unique species identifiers.}

\item{filters}{\code{data.frame}: generated by a call to \code{\link{select_filters}()}}

\item{locations}{\code{string}: generated by a call to \code{\link{select_locations}()}}

\item{columns}{\code{data.frame}: generated by a call to \code{\link{select_columns}()}}

\item{download_dir}{\code{string}: path to directory to store the downloaded media
in}
}
\value{
\code{data.frame} of media information if \code{metadata=TRUE}, otherwise
the number of images downloaded
}
\description{
In addition to text data describing individual occurrences and their attributes,
ALA stores image and sounds associated with a given record. These can be
downloaded to \code{R} by first using \code{\link{ala_occurrences}()} to
find records that contain media, and then passing the resulting \code{recordID}
field to \code{\link{ala_media}()} to download those media files (see examples).
}
\examples{
\dontrun{
# Download Regent Honeyeater multimedia
media_data <- ala_media(
    taxa = select_taxa("Regent Honeyeater"),
    filters = select_filters(year = 2020),
    download_dir = "media")

# Download just sounds of galahs
media_data <- ala_media(
     taxa = select_taxa("Eolophus Roseicapilla"),
     filters = select_filters(multimedia = "Sound"))
}
}
