% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_species.R
\name{ala_species}
\alias{ala_species}
\title{Species list}
\usage{
ala_species(taxa, filters, locations)
}
\arguments{
\item{taxa}{data.frame: generated by a call to \code{\link{select_taxa}}. This
argument also accepts a vector of unique species identifiers.}

\item{filters}{data.frame: generated by a call to \code{\link{select_filters}}}

\item{locations}{string: generated by a call to \code{\link{select_locations}}}
}
\value{
\code{data.frame} of matching species, with optional additional details.
}
\description{
Species list
}
