% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange.data_request}
\alias{arrange.data_request}
\alias{arrange.metadata_request}
\title{arrange rows}
\usage{
\method{arrange}{data_request}(.data, ...)

\method{arrange}{metadata_request}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{data_request}}

\item{...}{Either \code{count} or \code{index}}
}
\description{
New function to arrange on server side. Experimental
}
