% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_query.R
\name{as_query.data_request}
\alias{as_query.data_request}
\alias{as_query}
\alias{as_query.metadata_request}
\alias{as_query.files_request}
\title{Convert an object to class \code{query}}
\usage{
as_query(x, ...)

\method{as_query}{data_request}(x, ...)

\method{as_query}{metadata_request}(x, ...)

\method{as_query}{files_request}(x, thumbnail, ...)
}
\arguments{
\item{x}{An object to convert to a \code{query}. Supported classes are the same
as those produced by \code{\link[=galah_call]{galah_call()}}, namely \code{data_request},
\code{metadata_request} or \code{files_request}.}

\item{...}{Other arguments, currently ignored}

\item{thumbnail}{Logical: should thumbnail-size images be returned? Defaults
to \code{FALSE}, indicating full-size images are required.}
}
\value{
An object of class \code{query}, which is a list-like object containing at
least the slots \code{type} and \code{url}.
}
\description{
Functionally similar to \code{\link[=collapse]{collapse()}}, but without passing through
\code{\link[=coalesce]{coalesce()}} first. Primarily an internal function, but exported for
clarity and debugging purposes.
}
\details{
Typically, queries in galah are piped using \code{galah_call()}, which builds
an object of class \code{"data_request"}, \code{"metadata_request"} or \code{"files_request"}.
This parses to an object of class \code{"query"} via \code{\link[=collapse]{collapse()}}. However,
\code{\link[=collapse]{collapse()}} first calls \code{\link[=coalesce]{coalesce()}}, which expands to a \code{query_set}
\emph{before} evaluating \code{\link[=collapse]{collapse()}}. In this context, \code{as_query()} serves two
functions: externally, it can be called to convert directly to a \code{query}
without running checks; and internally it allows a query to be appended
to a \code{query_set} without calling \code{\link[=collapse]{collapse()}}, which would begin an
infinite loop (because \code{collapse()} calls \code{coalesce()}).

For simple cases, this gives the same result as running
\code{\link[=collapse]{collapse()}} while the \code{run_checks} argument of \code{\link[=galah_config]{galah_config()}} is set to
\code{FALSE}, but is slightly faster. For complex cases, however, it is likely
to generate irresolvable API calls, because e.g. taxonomic queries are not
parsed before the URL is built. It should therefore be used with care.
}
