% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_distributions.R
\name{atlas_distributions}
\alias{atlas_distributions}
\title{Collect a set of expert distributions}
\usage{
atlas_distributions(request = NULL, filter = NULL, identify = NULL)
}
\arguments{
\item{request}{optional \code{data_request} object: generated by a call to
\code{\link[=galah_call]{galah_call()}}.}

\item{filter}{\code{data.frame}: generated by a call to
\code{\link[=galah_filter]{galah_filter()}}}

\item{identify}{\code{data.frame}: generated by a call to
\code{\link[=galah_identify]{galah_identify()}}.}
}
\value{
An object of classes \code{sf}, \code{tbl}, \code{tbl_df} and \code{data.frame}
(aka a Simple feature collection) of distribution maps, with one column per
map, and spatial data stored in the \code{geometry} column.
}
\description{
Text goes here
}
\examples{

x <- show_all(distributions) |>
  slice_head(n = 4)

y <- galah_call() |>
  filter(id == x$id) |>
  atlas_distributions()
ggplot() +
  geom_sf(data = ozmap_country) +
  geom_sf(data = st_as_sf(x))
}
