% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_url.R
\name{atlas_url}
\alias{atlas_url}
\title{Internal function to build APIs}
\usage{
atlas_url(api_name, ..., quiet = FALSE, error_call = caller_env())
}
\description{
This function does a couple of things. Trivially, it looks up
base URLs from an internal tibble (\code{node_config}), itself imported from
a csv (\code{./data-raw/node_config.csv}). Entries in that tibble can contain
\code{glue}-like syntax for infilling new data, passed using \code{...}
}
\keyword{internal}
