% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_predicates.R
\name{build_predicates}
\alias{build_predicates}
\title{(currently internal) function to create 'predicates'}
\usage{
build_predicates(df)
}
\description{
(currently internal) function to create 'predicates'
}
