% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_cached_files.R
\name{clear_cached_files}
\alias{clear_cached_files}
\title{Clear previously cached files}
\usage{
clear_cached_files()
}
\description{
Deletes cached files within the cached file directory and their query
metadata
}
\section{Examples}{


First, set caching to true with \code{\link[=galah_config]{galah_config()}}. Then create a data query.
The data you download will be cached in a temporary directory.\if{html}{\out{<div class="sourceCode r">}}\preformatted{galah_config(caching = TRUE)
dat <- atlas_counts(group_by = galah_group_by(year))
}\if{html}{\out{</div>}}

To clear your cached files directory, use \code{clear_cached_files()}\if{html}{\out{<div class="sourceCode r">}}\preformatted{clear_cached_files()
}\if{html}{\out{</div>}}
}

