% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce.R
\name{coalesce}
\alias{coalesce}
\alias{coalesce.data_request}
\alias{coalesce.metadata_request}
\alias{coalesce.files_request}
\title{Force evaluation of a database query}
\usage{
coalesce(x, ...)

\method{coalesce}{data_request}(x, mint_doi, ...)

\method{coalesce}{metadata_request}(x, ...)

\method{coalesce}{files_request}(x, thumbnail, ...)
}
\arguments{
\item{x}{An object to be coalesced. Works for \code{data_request},
\code{metadata_request} and \code{file_request}.}

\item{...}{Other arguments}

\item{mint_doi}{Logical: should a DOI be minted for this download? Only
applies to \code{type = "occurrences"} when atlas chosen is "ALA".}

\item{thumbnail}{Logical: should thumbnail-size images be returned? Defaults
to \code{FALSE}, indicating full-size images are required.}
}
\value{
An object of class \code{query_set}, which is a list of all \code{query}
objects required to properly evaluate the specified request.
}
\description{
\code{coalesce} is an S3 generic function intended to be called before
\code{\link[=collapse]{collapse()}}. It is important as it shows the full set of queries
required to properly evaluate the user's request. This is often broader
than the single query returned by \code{\link[=collapse]{collapse()}}. It returns a \code{query_set}
object
}
