% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_occurrences.R
\name{collapse_occurrences}
\alias{collapse_occurrences}
\title{Internal function to \code{collapse()} for type = "occurrences"}
\usage{
collapse_occurrences(.data)
}
\description{
Internal function to \code{collapse()} for type = "occurrences"
}
