% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.data_query}
\alias{collect.data_query}
\title{Collect data from class \code{data_query}, returned by \code{collapse.data_request()}}
\usage{
\method{collect}{data_query}(.data)
}
\description{
Collect data from class \code{data_query}, returned by \code{collapse.data_request()}
}
