% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.data_response}
\alias{collect.data_response}
\title{Collect data from the selected atlas}
\usage{
\method{collect}{data_response}(x, wait = FALSE)
}
\arguments{
\item{x}{An object of class \code{data_response}, created using
\code{\link[=compute.data_request]{compute.data_request()}}}
}
\value{
A \code{tibble} containing requested data
}
\description{
In test
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\code{\link[=atlas_occurrences]{atlas_occurrences()}}
}
