% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_doi.R
\name{collect_doi}
\alias{collect_doi}
\title{Collect occurrence records from a pre-existing DOI}
\usage{
collect_doi(doi)
}
\arguments{
\item{doi}{\code{string}: this argument enables retrieval of occurrence
records previously downloaded from the ALA, using the DOI generated by the
data.}
}
\description{
This should be generated from \code{atlas_occurrences} or online
}
