% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_galah}
\alias{collect_galah}
\alias{collect.data_request}
\alias{collect.metadata_request}
\alias{collect.files_request}
\alias{collect.query_set}
\alias{collect.query}
\title{Retrieve a database query}
\usage{
\method{collect}{data_request}(.data, wait = TRUE, file = NULL)

\method{collect}{metadata_request}(.data)

\method{collect}{files_request}(.data)

\method{collect}{query_set}(.data, wait = TRUE, file = NULL)

\method{collect}{query}(.data, wait = TRUE, file = NULL)
}
\arguments{
\item{.data}{An object of class \code{data_request}, \code{metadata_request} or
\code{files_request} (from \code{galah_call()}); or an oject of class \code{query_set} or
\code{query} (from \code{collapse()} or \code{compute()})}

\item{wait}{logical; should \code{galah} wait for a response? Defaults to FALSE.
Only applies for \code{type = "occurrences"} or \code{"species"}.}

\item{file}{(optional) file name. If not given will be \code{data} with date and
time added. File path is always given by \code{galah_config()$package$directory}.}
}
\value{
In most cases, \code{collect()} returns a \code{tibble} containing requested
data. Where the requested data are not yet ready (i.e. for occurrences when
\code{wait} is set to \code{FALSE}), this function returns an object of class \code{query}
that can be used to recheck the download at a later time.
}
\description{
\code{collect()} attempts to retrieve the result of a query from the
selected API.
}
