% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_media.R
\name{collect_media}
\alias{collect_media}
\title{Collect media}
\usage{
collect_media(
  df,
  type = c("full", "thumbnail"),
  path,
  download_dir,
  refresh_cache
)
}
\arguments{
\item{df}{tibble returned by \code{atlas_media()} or \code{show_all_media()}}

\item{type}{\code{string}: either \code{"full"} to download original images, or
\code{"thumbnail"} to download thumbnails}

\item{path}{\code{string}: path to directory where downloaded media will
be stored}

\item{download_dir}{DEPRECATED: use \code{path} instead}

\item{refresh_cache}{\code{logical}: if set to \code{TRUE} and
\code{galah_config(caching = TRUE)} then files cached from a previous query will
be replaced by the current query. NOTE: unclear that this works right now.}
}
\description{
This function downloads images using information from \code{atlas_media}
}
