% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_occurrences.R
\name{collect_occurrences}
\alias{collect_occurrences}
\title{Collect occurrence records from a pre-existing DOI or URL}
\usage{
collect_occurrences(doi, url)
}
\arguments{
\item{doi}{\code{string}: Retrieve occurrence
records previously downloaded from the ALA, using the DOI generated by the
data.}

\item{url}{\code{string}: Retrieve occurrence
records previously downloaded from the ALA, using the URL provided via email.}
}
\description{
This should be generated from \code{atlas_occurrences} or online
}
