% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R, R/print.R
\name{compute.data_request}
\alias{compute.data_request}
\alias{print.data_response}
\title{Join the queue to download records from the chosen atlas}
\usage{
\method{compute}{data_request}(x)

\method{print}{data_response}(x, ...)
}
\arguments{
\item{x}{an object of class \code{data_response}}

\item{.data}{An object of class \code{data_request}, created using \code{\link[=galah_call]{galah_call()}}}
}
\value{
An object of class \code{request_info}
}
\description{
In test
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\code{\link[=atlas_occurrences]{atlas_occurrences()}}
}
