% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_request.R
\name{data_request}
\alias{data_request}
\title{Build a data request object}
\usage{
data_request(
  taxa = NULL,
  filters = NULL,
  locations = NULL,
  columns = NULL,
  ...
)
}
\arguments{
\item{taxa}{\code{data.frame}: generated by a call to \code{\link{select_taxa}()}. This
argument also accepts a vector of unique species identifiers.}

\item{filters}{\code{data.frame}: generated by a call to \code{\link{select_filters}()}}

\item{locations}{\code{string}: generated by a call to \code{\link{select_locations}()}}

\item{columns}{\code{data.frame}: generated by a call to \code{\link{select_columns}()}}

\item{...}{other function-specific request parameters}
}
\description{
Build a data request object
}
