% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{galah_down_to}
\alias{galah_down_to}
\title{Deprecated functions}
\usage{
galah_down_to(...)
}
\arguments{
\item{...}{the name of a single taxonomic rank}
}
\value{
A string with the named rank
}
\description{
These include:
\itemize{
\item \code{galah_down_to()} in favour of \code{galah_filter()}
}
}
\examples{
\dontrun{
# Return a taxonomic tree of *Chordata* down to the class level
galah_call() |>
    galah_identify("Vertebrata") |>
    galah_down_to(class) |>
    atlas_taxonomy()
}
}
\seealso{
\code{\link[=galah_select]{galah_select()}}, \code{\link[=galah_filter]{galah_filter()}} and
\code{\link[=galah_geolocate]{galah_geolocate()}} for related methods.
}
