% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{deprecated}
\alias{deprecated}
\alias{select_columns}
\alias{select_filters}
\alias{select_locations}
\alias{ala_occurrences}
\alias{ala_counts}
\alias{ala_species}
\alias{ala_taxonomy}
\alias{ala_media}
\alias{ala_citation}
\title{Deprecated functions}
\usage{
select_columns(..., group)

select_filters(..., profile = NULL)

select_locations(query)

ala_occurrences(
  taxa = NULL,
  filters = NULL,
  locations = NULL,
  columns = select_columns(group = "basic"),
  mint_doi = FALSE,
  doi,
  refresh_cache = FALSE
)

ala_counts(
  taxa = NULL,
  filters = NULL,
  locations = NULL,
  limit = 100,
  type = c("record", "species"),
  refresh_cache = FALSE
)

ala_species(
  taxa = NULL,
  filters = NULL,
  locations = NULL,
  refresh_cache = FALSE
)

ala_taxonomy(taxa, down_to)

ala_media(
  taxa = NULL,
  filters = NULL,
  locations = NULL,
  columns = select_columns(group = "basic"),
  download_dir,
  refresh_cache = FALSE
)

ala_citation(data)
}
\description{
`r lifecycle::badge("deprecated")`

The following functions are deprecated as of `galah` version 2.0.0.

Use [galah_select()] instead of `select_columns()`.

Use [galah_filter()] instead of `select_filters()`.

Use [galah_location()] instead of `select_locations()`.

Use [atlas_counts()] instead of `ala_counts()`.

Use [atlas_species()] instead of `ala_species()`.

Use [atlas_taxonomy()] instead of `ala_taxonomy()`.

Use [atlas_media()] instead of `ala_media()`.

Use [atlas_citation()] instead of `ala_citation()`.
}
