% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-desc.R
\name{desc}
\alias{desc}
\title{Descending order}
\usage{
desc(...)
}
\arguments{
\item{...}{column to order by}
}
\value{
A \code{tibble} used by \code{arrange.data_request()} to arrange rows of a
query.
}
\description{
Descending order
}
\examples{
\dontrun{
# Arrange grouped record counts by descending year
galah_call() |>
  identify("perameles") |>
  filter(year > 2019) |>
  count() |>
  arrange(galah::desc(year)) |>
  collect()
}
}
\seealso{
\code{\link[=arrange.data_request]{arrange()}}, \code{\link[=galah_call]{galah_call()}}
}
