% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_filters.R
\name{exclude}
\alias{exclude}
\title{Negate a filter value}
\usage{
exclude(value)
}
\arguments{
\item{value}{string: filter value(s) to be excluded}
}
\value{
value with class "exclude"
}
\description{
Deprecated alternative to \code{select_filters(field != value)}.
}
\seealso{
\code{exclude} is used with \code{\link{select_filters}} or
\code{\link{select_taxa}} to exclude values
}
