% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_config.R
\name{find_atlases}
\alias{find_atlases}
\title{List supported Living Atlases}
\usage{
find_atlases()
}
\value{
a character vector of supported Living Atlases
}
\description{
galah supports downloading data from a number of International Living
Atlases. Use this function to get a list of all currently supported atlases.
}
\seealso{
This function is helpful in setting up \code{\link{ala_config}()}.
}
