% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cached_files.R
\name{find_cached_files}
\alias{find_cached_files}
\title{List previously cached files}
\usage{
find_cached_files()
}
\value{
a \code{list} of available cached files, the function used to
create them, and the filter object
}
\description{
Uses query metadata stored in metadata.rds in the cache directory
}
\examples{
## configure caching and create a query to cache
\dontrun{
galah_config(caching = TRUE)
dat <- ala_counts(group_by = "year")

## list cached files
find_cached_files()
}
}
