% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_layers.R
\name{find_layers}
\alias{find_layers}
\title{Search ALA spatial layers}
\usage{
find_layers()
}
\value{
A \code{data.frame} of all spatial layers held by the ALA
}
\description{
To return layer values with occurrence records, pass the required layer
names to \code{\link{select_columns}}.
}
\examples{
\dontrun{
# Find all precipitation-related layers
layers <- find_layers()
layers[grepl("Precipitation", layers$name),]
}
}
\seealso{
This function is used to pass valid arguments to
\code{\link{select_columns}}. Layers are the subset of fields that are
spatially appended to each record by the ALA; the full list is given by
\code{\link{find_fields}}.
}
