% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_profiles.R
\name{find_profile_attributes}
\alias{find_profile_attributes}
\title{Get data filters for a  data quality profile}
\usage{
find_profile_attributes(profile)
}
\arguments{
\item{profile}{string: a data quality profile name, short name or id.
See `find_profiles` for valid filters}
}
\value{
data.frame of profile attributes
}
\description{
Get data filters for a  data quality profile
}
\examples{
find_profile_attributes("CSDM")
}
