% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_data_profiles.R
\name{find_profile_filters}
\alias{find_profile_filters}
\title{Get data filters for a  data quality profile}
\usage{
find_profile_filters(profile)
}
\arguments{
\item{profile}{string: a data quality profile name, short name or id.
See `find_data_profiles` for valid filters}
}
\description{
Get data filters for a  data quality profile
}
