% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_profiles.R
\name{find_profiles}
\alias{find_profiles}
\title{Data quality profiles}
\usage{
find_profiles()
}
\value{
A \code{data.frame} of available profiles
}
\description{
List data quality profiles that can be used for data filtering.
}
\examples{
# Get available profiles
profile_df <- find_profiles()
# Values given in the 'shortName' column are accepted by select_filter(), i.e.
select_filter(profle = profile_df$shortName[1])
is equivalent to:
select_filter(profle = "ALA")
}
\seealso{
This function gives viable profle names for passing to
\code{\link{select_filters}}. For more detail on a given profile see
\code{\link{find_profile_attributes}}.
}
