% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_ranks.R
\name{find_ranks}
\alias{find_ranks}
\title{Find valid taxonomic ranks}
\usage{
find_ranks()
}
\value{
A \code{data.frame} of available ranks
}
\description{
Return taxonomic ranks recognised by the ALA.
}
\examples{
\dontrun{
rank_df <- find_ranks()
}
}
\seealso{
This function provides a
reference that is useful when specifying the \code{down_to} argument of 
\code{\link{ala_taxonomy}}.
}
