% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_config.R
\name{find_reasons}
\alias{find_reasons}
\title{List valid download reasons}
\usage{
find_reasons()
}
\description{
List valid download reasons
}
\seealso{
This function is helpful in setting up \code{\link{ala_config}}.
}
