% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah-package.R
\docType{package}
\name{galah-package}
\alias{galah}
\alias{galah-package}
\title{galah: Atlas of Living Australia (ALA) Data and Resources in R}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The Atlas of Living Australia ('ALA') provides tools to enable users of biodiversity information to find, access, combine and visualise data on Australian plants and animals; these have been made available from <https://ala.org.au/>. 'galah' provides a subset of the tools to be directly used within R. It enables the R community to directly access data and resources hosted by the 'ALA'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AtlasOfLivingAustralia/galah}
  \item Report bugs at \url{https://github.com/AtlasOfLivingAustralia/galah/issues}
}

}
\author{
\strong{Maintainer}: Martin Westgate \email{martin.westgate@csiro.au}

Authors:
\itemize{
  \item Matilda Stevenson \email{matilda.stevenson@csiro.au}
  \item Dax Kellie \email{dax.kellie@csiro.au}
  \item Peggy Newman \email{peggy.newman@csiro.au}
}

}
\keyword{internal}
