% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah_data_profile.R
\name{galah_data_profile}
\alias{galah_data_profile}
\title{Specify a data profile}
\usage{
galah_data_profile(...)
}
\arguments{
\item{...}{a profile name. Valid values can be seen using \code{show_all_profiles()$shortName}}
}
\description{
'profiles' are groups of filters that are pre-applied by the ALA, allowing
rapid narrowing of results to the most relevant data. Note that only one
profile can be loaded at a time; if multiple profiles are given, the first
valid profile is used.
}
