% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah_down_to.R
\name{galah_down_to}
\alias{galah_down_to}
\title{Specify the lowest taxonomic rank required in a downwards search}
\usage{
galah_down_to(...)
}
\arguments{
\item{...}{the name of a single taxonomic rank}
}
\value{
A string with the named rank
}
\description{
\code{atlas_taxonomy} generates a downwards search of the taxonomic tree. This
function can be used to specify the name of a valid taxonomic rank using
non-standard evaluation (NSE), for consistency with other \code{galah_} functions.
}
\examples{
# A basic case
atlas_taxonomy(
    taxa = search_taxa("Chordata")
    down_to = galah_down_to(class)
    )
}
\seealso{
\code{\link[=galah_select]{galah_select()}}, \code{\link[=galah_filter]{galah_filter()}} and
\code{\link[=galah_geolocate]{galah_geolocate()}} for related methods.
}
