% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah_identify.R
\name{galah_identify}
\alias{galah_identify}
\title{Build a query that contains taxonomic identifiers}
\usage{
galah_identify(..., search = TRUE)
}
\arguments{
\item{...}{an object of class \code{ala_id}, \code{gbifid}, \code{nbnid} or \code{character}.}

\item{search}{(logical); should the results in question be passed to \code{search_taxa}?}
}
\description{
A common task is to restrict searches to known taxonomic groups. This function
allows users to pass identifiers found using \code{search_taxa} within a pipe to
provide data only for the biological group of interest.
}
\seealso{
\code{\link[=search_taxa]{search_taxa()}} to find identifiers from scientific names;
\code{\link[=search_identifiers]{search_identifiers()}} for how to get names if taxonomic identifiers
are already known.
}
