% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah_location.R
\name{galah_location}
\alias{galah_location}
\title{Build a WKT string from an sf spatial object or verify an existing WKT}
\usage{
galah_location(query)
}
\arguments{
\item{query}{wkt string or sf object}
}
\value{
length-1 object of class \code{character} and \code{ala_locations},
containing a WKT string representing the area provided.
}
\description{
Restrict results to those from a specified area. Areas must be polygons
and be specified as either an sf object, or a 'well-known text' (wkt) string.
}
\details{
WKT strings longer than 10000 characters will not be
accepted by the ALA- so the sf object or WKT string may need to be
simplified.
}
\examples{
\dontrun{
# Search for records using a shapefile
locations <- galah_location(st_read(path/to/shapefile))
ala_occurrences(locations = locations)

# Search for records using a WKT
wkt <- "POLYGON((142.36228 -29.00703,142.74131 -29.00703,142.74131 \
-29.39064,142.36228 -29.39064,142.36228 -29.00703))"
ala_occurrences(locations = galah_location(wkt))
}
}
\seealso{
\code{\link{select_taxa}}, \code{\link{galah_filter}} and
\code{\link{select_columns}} for other ways to restrict the information
returned by \code{\link{ala_occurrences}} and related functions.
}
