% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah_media.R
\name{galah_media}
\alias{galah_media}
\alias{media}
\title{Function to specify how media should be handled by \code{galah}}
\usage{
galah_media(
  .data,
  path = ".",
  filesize = c("full", "thumbnail"),
  media = c("image", "sound", "video")
)

media(.data, path, filesize, media)
}
\arguments{
\item{.data}{An object of class \code{data_request}}

\item{path}{String: where should data be saved?}

\item{filesize}{String: either "full" (default) or "thumbnail". Note only
applies if \code{media = "image"}.}

\item{media}{Vector: One or more of "image", "sound" or "video", defaulting
to all three.}
}
\description{
In prep
}
