% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koala.R
\docType{package}
\name{koala}
\alias{koala}
\title{\pkg{koala}}
\description{
This project enables the R community to access data and tools hosted by the
Atlas of Living Australia. The goal of the project is
to enable basic species and related information to be queried and output in
standard formats for R. koala is based around the extensive web services
provided by the Atlas; see the API link below.
}
\details{
If you have any questions or suggestions, please email support@ala.org.au.
}
\references{
\url{https://api.ala.org.au/}
}
