% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{request_data}
\alias{request_data}
\alias{request_metadata}
\title{new functions to start pipes for different types of data}
\usage{
request_data(
  type = c("occurrences", "species", "occurrences-count", "species-count"),
  ...
)

request_metadata(
  type = c("fields", "apis", "assertions", "atlases", "collections", "datasets",
    "layers", "licences", "lists", "profiles", "providers", "ranks", "reasons")
)
}
\value{
Each function returns a different object class: \code{request_data()}
returns \code{data_request} (synonymous with \code{galah_call()}). \code{request_metadata}
returns \code{metadata_request}.
\code{download_request}? \code{file_request}? \code{media_request}? Not sure on last type
}
\description{
Each function can begin a pipe and end with \code{collapse()}, \code{compute()} or
\code{collect()}. Having a different piping function is useful because it allows
later functions to have class-specific requests; e.g. \code{filter.data_request}
generates filters for \code{solr}, whereas \code{filter.metadata_request} is a search
function that accepts a single string.
}
\details{
More pragmatically, adding \code{request_metadata} enables use of \code{collect()} etc
for metadata functions, which make it easier to build and debug these APIs.
}
