% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fields.R
\name{search_fields}
\alias{search_fields}
\title{Query layers or fields by free text search}
\usage{
search_fields(query, type = "both")
}
\arguments{
\item{query}{\code{string}: A search string. Not case sensitive.}

\item{type}{\code{string}: What type of parameters should be searched?
Should be one of <code>fields</code>, <code>layers</code> or <code>both</code>.}
}
\value{
A <code>data.frame</code> with three columns:
\itemize{
 \item{id: The identifier for that layer or field. This is the value that should
 be used when referring to a field in another function.}
 \item{description: Detailed information on a given field}
 \item{type: Whether the field is a <code>field</code> or <code>layer</code>}
}
}
\description{
Function to search the ALA database for layers or fields
using a free-text search.
}
\examples{
\dontrun{
test <- search_fields("species")
}
}
