% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fields.R, R/search_identifiers.R,
%   R/search_minifunctions.R, R/search_profile_attributes.R, R/search_taxa.R,
%   R/search_values.R, R/show_all_assertions.R, R/show_all_atlases.R,
%   R/show_all_collectory.R, R/show_all_profiles.R, R/show_all_ranks.R,
%   R/show_all_reasons.R, R/show_all_values.R
\name{search_fields}
\alias{search_fields}
\alias{search_identifiers}
\alias{search_minifunctions}
\alias{search_atlases}
\alias{search_reasons}
\alias{search_taxa}
\alias{search_ranks}
\alias{search_values}
\alias{search_assertions}
\alias{search_profiles}
\alias{search_profile_attributes}
\alias{search_providers}
\alias{search_collections}
\alias{search_datasets}
\title{Functions for searching detailed metadata}
\usage{
search_fields(query)

search_identifiers(identifier)

search_profile_attributes(profile)

search_taxa(...)

search_values(query, field)

search_assertions(query)

search_atlases(query)

search_collections(query)

search_datasets(query)

search_providers(query)

search_profiles(query)

search_ranks(query)

search_reasons(query)

search_values(query, field)
}
\arguments{
\item{query}{\code{string}: A search string. Not case sensitive.}

\item{identifier}{\code{string}: A vector containing one or more taxonomic
identifiers, given as strings.}

\item{profile}{\code{string}: a data quality profile name, short name or id.
See \code{\link[=show_all_profiles]{show_all_profiles()}} for valid filters}

\item{...}{: One or more scientific names, separated by commas and
given as strings. If greater control is required to disambiguate search
terms, taxonomic levels can be provided explicitly via a \code{data.frame}
(see examples). Note that searches are not case-sensitive.}

\item{field}{\code{string}: field to return the categories for. Use
\code{\link[=search_fields]{search_fields()}} to view valid fields.}
}
\description{
Some descriptive text
}
\section{Examples}{


Search using a single term\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_taxa("Reptilia")
#> # A tibble: 1 x 9
#>   search_term scientific_name taxon_concept_id                        rank  match_type kingdom  phylum class  issues
#>   <chr>       <chr>           <chr>                                   <chr> <chr>      <chr>    <chr>  <chr>  <chr> 
#> 1 Reptilia    REPTILIA        urn:lsid:biodiversity.org.au:afd.taxon~ class exactMatch Animalia Chord~ Repti~ noIss~
}\if{html}{\out{</div>}}

Note that \code{search_taxa()} is not case sensitive\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_taxa("reptilia") # not case sensitive
#> # A tibble: 1 x 9
#>   search_term scientific_name taxon_concept_id                        rank  match_type kingdom  phylum class  issues
#>   <chr>       <chr>           <chr>                                   <chr> <chr>      <chr>    <chr>  <chr>  <chr> 
#> 1 reptilia    REPTILIA        urn:lsid:biodiversity.org.au:afd.taxon~ class exactMatch Animalia Chord~ Repti~ noIss~
}\if{html}{\out{</div>}}

Search multiple taxa. \code{search_taxa()} will return one row per taxon\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_taxa(c("reptilia", "mammalia"))
#> # A tibble: 2 x 10
#>   search_term scientific_name taxon_concept_id          rank  match_type kingdom phylum class issues vernacular_name
#>   <chr>       <chr>           <chr>                     <chr> <chr>      <chr>   <chr>  <chr> <chr>  <chr>          
#> 1 reptilia    REPTILIA        urn:lsid:biodiversity.or~ class exactMatch Animal~ Chord~ Rept~ noIss~ <NA>           
#> 2 mammalia    MAMMALIA        urn:lsid:biodiversity.or~ class exactMatch Animal~ Chord~ Mamm~ noIss~ Mammals
}\if{html}{\out{</div>}}

\code{galah_identify()} uses \code{search_taxa()} to narrow data queries

Look up a unique taxon identifier\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_identifiers(identifier = "https://id.biodiversity.org.au/node/apni/2914510")
#> # A tibble: 1 x 13
#>   scientific_name scientific_name~ taxon_concept_id rank  match_type kingdom phylum class order family genus species
#>   <chr>           <chr>            <chr>            <chr> <chr>      <chr>   <chr>  <chr> <chr> <chr>  <chr> <chr>  
#> 1 Eucalyptus bla~ Maiden           https://id.biod~ spec~ taxonIdMa~ Plantae Charo~ Equi~ Myrt~ Myrta~ Euca~ Eucaly~
#> # ... with 1 more variable: issues <chr>
}\if{html}{\out{</div>}}\if{html}{\out{<div class="sourceCode r">}}\preformatted{galah_call() |>
  galah_identify("reptilia") |>
  atlas_counts()
#> # A tibble: 1 x 1
#>     count
#>     <int>
#> 1 1322791
}\if{html}{\out{</div>}}

Search for all fields that use include the word "date"\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("date")
#> # A tibble: 33 x 4
#>   id             description     type   link 
#>   <chr>          <chr>           <chr>  <chr>
#> 1 dateIdentified Date Identified fields <NA> 
#> 2 datePrecision  Date precision  fields <NA> 
#> 3 eventDate      Event Date      fields <NA> 
#> 4 eventDateEnd   <NA>            fields <NA> 
#> # ... with 29 more rows
}\if{html}{\out{</div>}}

Search for all fields with the string "basisofrecord"\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("basisofrecord")
#> # A tibble: 2 x 4
#>   id                description               type   link 
#>   <chr>             <chr>                     <chr>  <chr>
#> 1 basisOfRecord     Record type               fields <NA> 
#> 2 raw_basisOfRecord Record type (unprocessed) fields <NA>
}\if{html}{\out{</div>}}

Search for all fields that have information for "marine"\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("marine") |> 
  head() # only show first 5 results
#> # A tibble: 6 x 4
#>   id      description                                          type  link                                           
#>   <chr>   <chr>                                                <chr> <chr>                                          
#> 1 cl2105  "FAO Fishery Statistical Areas FAO Fishing areas (M~ laye~ "http://www.fao.org/geonetwork/srv/en/main.hom~
#> 2 cl10948 "CAPAD 2018 Terrestrial The Collaborative Australia~ laye~ "http://www.environment.gov.au/fed/catalog/sea~
#> 3 cl10957 "CAPAD 2018 Marine The Collaborative Australian Pro~ laye~ "http://www.environment.gov.au/fed/catalog/sea~
#> 4 el1056  "Endemism (Non-marine) Endemism (Non-marine) based ~ laye~ ""                                             
#> 5 el957   "Averaged Topographic Relief  This data represents ~ laye~ "http://www.ga.gov.au/meta/ANZCW0703015076.htm~
#> 6 cl11033 "CAPAD 2020 Terrestrial The Collaborative Australia~ laye~ "http://www.environment.gov.au/fed/catalog/sea~
}\if{html}{\out{</div>}}

Search for all Wordclim layers\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("worldclim")
#> # A tibble: 38 x 4
#>   id      description                                                                     type   link               
#>   <chr>   <chr>                                                                           <chr>  <chr>              
#> 1 el10982 WorldClim 2.1: Temperature - warmest month max Max Temperature of Warmest Month layers https://www.worldc~
#> 2 el10981 WorldClim 2.1: Temperature - seasonality Temperature Seasonality                layers https://www.worldc~
#> 3 el10980 WorldClim 2.1: Temperature - isothermality Isothermality                        layers https://www.worldc~
#> 4 el10990 WorldClim 2.1: Precipitation - wettest month Precipitation of Wettest Month     layers https://www.worldc~
#> # ... with 34 more rows
}\if{html}{\out{</div>}}

To find all the data quality arguments used in the profile "CSDM"\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_profile_attributes("CSDM")
#> # A tibble: 4 x 2
#>   description                                                               filter                       
#>   <chr>                                                                     <chr>                        
#> 1 "Include only records where Spatial validity is \\"true\\""                 "spatiallyValid:\\"true\\""    
#> 2 "Exclude potential duplicate records"                                     "-isDuplicateOf:*"           
#> 3 "Exclude all records that are an outlier against any environmental layer" "-outlierLayerCount:[1 TO *]"
#> 4 "Include only records where Year is 1970 to 2099"                         "year:[1970 TO *]"
}\if{html}{\out{</div>}}

Then get a free-text description of each filter used in the "CSDM" profile\if{html}{\out{<div class="sourceCode r">}}\preformatted{profile_info <- search_profile_attributes("CSDM")
profile_info$description
#> [1] "Include only records where Spatial validity is \\"true\\""                
#> [2] "Exclude potential duplicate records"                                    
#> [3] "Exclude all records that are an outlier against any environmental layer"
#> [4] "Include only records where Year is 1970 to 2099"
}\if{html}{\out{</div>}}
}

