% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_values.R
\name{search_values}
\alias{search_values}
\title{Search for a value within a field}
\usage{
search_values(field, query)
}
\arguments{
\item{field}{\code{string}: field to return the categories for. Use
\code{\link[=search_fields]{search_fields()}} to view valid fields.}

\item{query}{\code{string}: A search string. Not case sensitive.}
}
\description{
Descriptive text goes here.
}
