% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_columns.R
\name{select_columns}
\alias{select_columns}
\title{Specify columns for occurrence download}
\usage{
select_columns(..., group)
}
\arguments{
\item{...}{zero or more individual column names to include}

\item{group}{string: (optional) name of one or more column groups to
include. Options are \code{'basic'}, \code{'event'}) and
\code{'assertion'}}
}
\description{
The result of \code{select_columns} can be passed to the \code{columns}
argument in \code{\link{ala_occurrences}}.
}
\details{
Calling the argument \code{group = 'basic'} returns the following columns:
\itemize{
  \item\code{decimalLatitude}
  \item\code{decimalLongitude}
  \item\code{eventDate}
  \item\code{taxon_name}
  \item\code{taxonConceptID}
  \item\code{recordID}
  \item\code{data_resource}
}
Using \code{group = 'event'} returns the following columns:
\itemize{
  \item\code{eventRemarks}
  \item\code{eventTime}
  \item\code{eventID}
  \item\code{eventDate}
  \item\code{samplingEffort}
  \item\code{samplingProtocol}
}
Using \code{group = 'assertion'} returns all quality assertion-related
columns. The list of assertions is shown by `find_fields("assertion")`.
}
