% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_fields.R
\name{select_fields}
\alias{select_fields}
\alias{select_columns}
\title{Specify fields for occurrence download}
\usage{
select_fields(..., group = c("basic", "event", "assertions"), expand = FALSE)

select_columns(..., group, expand)
}
\arguments{
\item{...}{zero or more individual column names to include}

\item{group}{\code{string}: (optional) name of one or more column groups to
include. Valid options are \code{"basic"}, \code{"event"} and
\code{"assertion"}}

\item{expand}{\code{logical}: When passed to \code{group_by} argument of 
\code{ala_counts}, should factor levels be expanded? Defaults to \code{FALSE}.}
}
\value{
An object of class \code{data.frame} and \code{ala_fields}
specifying the name and type of each column to include in the 
call to \code{ala_counts} or |code{ala_occurrences}.
}
\description{
The ALA stores content on hundreds of different fields, and users often
thousands or millions of records at a time. To reduce time taken to download
data, and limit complexity of the resulting \code{data.frame}, it is often
sensible to restrict the fields returned by \code{\link{ala_occurrences}()}
or when using the \code{group_by} argument of \code{\link{ala_counts}}.
This function allows easy selection of fields, or commonly-requested groups 
of columns.
}
\details{
Calling the argument \code{group = "basic"} returns the following columns:
\itemize{
  \item\code{decimalLatitude}
  \item\code{decimalLongitude}
  \item\code{eventDate}
  \item\code{scientificName}
  \item\code{taxonConceptID}
  \item\code{recordID}
  \item\code{dataResourceName}
}
Using \code{group = "event"} returns the following columns:
\itemize{
  \item\code{eventRemarks}
  \item\code{eventTime}
  \item\code{eventID}
  \item\code{eventDate}
  \item\code{samplingEffort}
  \item\code{samplingProtocol}
}
Using \code{group = "assertions"} returns all quality assertion-related
columns. The list of assertions is shown by \code{search_fields(type = "assertions")}.

The \code{expand} argument is appended as an attribute of the \code{data.frame} 
returned by \code{select_fields}. It is only used by \code{ala_counts}, meaning
that setting \code{expand = TRUE} will not affect calls to \code{ala_occurrences}.
}
\note{
\code{select_columns} and \code{select_fields} are synonymous;
\code{select_columns} is deprecated and will be removed from future versions
of \code{galah}.
}
\seealso{
\code{\link{select_taxa}}, \code{\link{select_filters}} and
\code{\link{select_locations}} for other ways to restrict the information returned
by \code{\link{ala_occurrences}} and related functions; \code{\link{ala_counts}}
for how to get counts by levels of variables returned by \code{select_fields}.
}
