% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_filters.R
\name{select_filters}
\alias{select_filters}
\title{Select filters to narrow down occurrence queries}
\usage{
select_filters(..., profile = NULL)
}
\arguments{
\item{...}{filters, in the form field = value}

\item{profile}{string: (optional) a data quality profile to apply to the
records. See \code{\link{find_profiles}} for valid profiles. By default
no profile is applied.}
}
\value{
dataframe of filter values
}
\description{
The result of `select_filters` can be passed to the `filters` argument in
\code{\link{ala_occurrences}} and \code{\link{ala_counts}}
}
