% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_groups.R
\name{select_groups}
\alias{select_groups}
\title{Specify groups for which counts are required}
\usage{
select_groups(..., expand = FALSE)
}
\arguments{
\item{...}{one or more strings giving variable names (see \code{\link{search_fields}})}

\item{expand}{Logical: should all combinations of field levels be returned?}
}
\value{
An object of class \code{data.frame} and \code{ala_groups}
containing columns \code{name} (string) and \code{expand} (logical).
}
\description{
A function to specify how counts should be grouped using 
\code{\link{ala_counts}}.
}
\details{
If \code{expand} is \code{FALSE} (default) and >1 variables are specified
using \code{...}, then the counts for those variables are calculated separately.
If \code{TRUE}, then all n-wise combinations are calculated, where n is the
number of variables supplied.
}
