% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all.R
\name{show_all}
\alias{show_all}
\alias{search_all}
\title{Look up record metadata}
\usage{
show_all(type, field)

search_all(type, query)
}
\arguments{
\item{type}{\code{string}: What type of parameters should be searched?
See \code{Details} (below) for accepted values.}

\item{query}{\code{string}: A search string. Not case sensitive.}
}
\value{
An object of class \code{tbl_df} and \code{data.frame} (aka a tibble)
containing all data of interest (for \code{show_all()}), or those that match the
search query (for \code{search_all()}).
}
\description{
The living atlases store a huge amount of information, above and beyond the
occurrence records that are their main output. In \code{galah}, users can investigate
the metadata needed to build complex queries using two meta-functions or a larger
number of sub-functions.
}
\details{
There are five different categories of metadata, each with their own
sub-functions. Valid \code{type}s are:\tabular{llll}{
   \strong{Category} \tab \strong{Type} \tab \strong{Description} \tab \strong{Sub-functions} \cr
   configuration \tab \verb{atlases"} \tab Show what atlases are available \tab \code{\link[=show_all_atlases]{show_all_atlases()}}, \code{\link[=search_atlases]{search_atlases()}} \cr
    \tab \code{apis} \tab Show what APIs & functions are available for each atlas \tab \code{\link[=show_all_apis]{show_all_apis()}}, \code{\link[=search_apis]{search_apis()}} \cr
    \tab \code{reasons} \tab Show what values are acceptable as 'download reasons' for a specified atlas \tab \code{\link[=show_all_reasons]{show_all_reasons()}}, \code{\link[=search_reasons]{search_reasons()}} \cr
   taxonomy \tab \code{taxa} \tab Search for one or more taxonomic names \tab \code{\link[=search_taxa]{search_taxa()}} \cr
    \tab \code{identifiers} \tab Take a universal identifier and return taxonomic information \tab \code{\link[=search_identifiers]{search_identifiers()}} \cr
    \tab \code{ranks} \tab Show valid taxonomic ranks (e.g. Kingdom, Class, Order, etc.) \tab \code{\link[=show_all_ranks]{show_all_ranks()}}, \code{\link[=search_ranks]{search_ranks()}}) \cr
   filters \tab \code{fields} \tab Show fields that are stored in an atlas \tab \code{\link[=show_all_fields]{show_all_fields()}}, \code{\link[=search_fields]{search_fields()}} \cr
    \tab \code{assertions} \tab Show results of data quality checks run by each atlas \tab \code{\link[=show_all_assertions]{show_all_assertions()}}, \code{\link[=search_assertions]{search_assertions()}} \cr
   group filters \tab \code{profiles} \tab Show what data profiles are available \tab \code{\link[=show_all_profiles]{show_all_profiles()}}, \code{\link[=search_profiles]{search_profiles()}} \cr
    \tab \code{lists} \tab Show what species lists are available \tab  \cr
   data providers \tab \code{providers} \tab Show which institutions have provided data \tab \code{\link[=show_all_providers]{show_all_providers()}}, \code{\link[=search_providers]{search_providers()}} \cr
    \tab \code{collections} \tab Show the specific collections within those institutions \tab \code{\link[=show_all_collections]{show_all_collections()}}, \code{\link[=search_collections]{search_collections()}} \cr
    \tab \code{datasets} \tab Shows all the data groupings within those collections \tab \code{\link[=show_all_datasets]{show_all_datasets()}}, \code{\link[=search_datasets]{search_datasets()}} \cr
}


Each valid \code{type} within \code{show_all} has a corresponding internal function where
type is used as a suffix. For example, \code{show_all(fields)} is
synonymous with \code{show_all_fields()}, while \code{search_all(fields, "basisOfRecord")}
is synonymous with \code{search_fields("basisOfRecord")}
}
\references{
\itemize{
\item Darwin Core terms \url{https://dwc.tdwg.org/terms/}
\item ALA fields \url{https://api.ala.org.au/#ws72}
\item ALA assertions fields \url{https://api.ala.org.au/#ws81}
}
}
\seealso{
These functions are used to pass valid arguments to
\code{\link[=galah_select]{galah_select()}}, \code{\link[=galah_filter]{galah_filter()}}, and related functions.
}
