% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_atlases.R
\name{show_all_atlases}
\alias{show_all_atlases}
\title{List supported Living Atlases}
\usage{
show_all_atlases()
}
\value{
a \code{data.frame} of Living Atlas information, including taxonomy
source and information for each atlas.
}
\description{
galah supports downloading data from a number of International Living
Atlases. Use this function to get a list of all currently supported atlases.
}
\seealso{
This function is helpful in setting up \code{\link{galah_config}()}.
}
