% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fields.R, R/show_all_fields.R
\name{search_fields}
\alias{search_fields}
\alias{show_all_fields}
\title{Query layers, fields or assertions by free text search}
\usage{
search_fields(
  query,
  type = c("all", "fields", "layers", "assertions", "media", "other")
)

show_all_fields(
  type = c("all", "fields", "layers", "assertions", "media", "other")
)
}
\arguments{
\item{query}{\code{string}: A search string. Not case sensitive.}

\item{type}{\code{string}: What type of parameters should be searched?
Should be one or more of \code{fields}, \code{layers}, \code{assertions},
\code{media} or \code{all}.}
}
\value{
if \code{query} is missing, an empty \code{data.frame}; otherwise
an object of class \code{tbl_df} and \code{data.frame} (aka a tibble) containing
fields that match the search query.

An object of class \code{tbl_df} and \code{data.frame} (aka a tibble) with
three columns:
\itemize{
\item id: The identifier for that layer or field. This is the value that should
be used when referring to a field in another function.
\item description: Detailed information on a given field
\item type: Whether the field is a \code{field} or \code{layer}
\item link: For layers, a link to the source data (if available)
}
}
\description{
This function can be used to find relevant fields and/or layers
for use in building a set of filters with \code{\link[=galah_filter]{galah_filter()}} or
specifying required columns with \code{\link[=galah_select]{galah_select()}}.
This function returns a \code{data.frame} of all fields matching the type
specified.
Field names are in Darwin Core format, except in the case where the field is
specific to the ALA database, in which case the ALA field name is returned.
}
\details{
Layers are the subset of fields that are spatially appended to each record
by the ALA. Layer ids are comprised of a prefix: 'el' for environmental
(gridded) layers and 'cl' for contextual (polygon) layers,  followed by an
id number.
}
\section{Examples}{


Search for all fields that use include the word "date"\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("date")
#> # A tibble: 32 × 4
#>    id                        description                                                                    type  link 
#>    <chr>                     <chr>                                                                          <chr> <chr>
#>  1 dateIdentified            http://rs.tdwg.org/dwc/terms/dateIdentified                                    fiel… <NA> 
#>  2 datePrecision             The precision of the date information for the record. Values include Day, Mon… fiel… <NA> 
#>  3 eventDate                 The ISO formatted date of observation, specimen collection date. http://rs.td… fiel… <NA> 
#>  4 eventDateEnd              <NA>                                                                           fiel… <NA> 
#>  5 firstLoadedDate           <NA>                                                                           fiel… <NA> 
#>  6 georeferencedDate         http://rs.tdwg.org/dwc/terms/georeferencedDate                                 fiel… <NA> 
#>  7 lastAssertionDate         <NA>                                                                           fiel… <NA> 
#>  8 lastLoadDate              <NA>                                                                           fiel… <NA> 
#>  9 lastProcessedDate         <NA>                                                                           fiel… <NA> 
#> 10 measurementDeterminedDate http://rs.tdwg.org/dwc/terms/measurementDeterminedDate                         fiel… <NA> 
#> # … with 22 more rows
}\if{html}{\out{</div>}}

Search for all fields with the string "basisofrecord"\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("basisofrecord")
#> # A tibble: 2 × 4
#>   id                description                                                                             type  link 
#>   <chr>             <chr>                                                                                   <chr> <chr>
#> 1 basisOfRecord     What this is a record of e.g. specimen, human observation, fossil http://rs.tdwg.org/d… fiel… <NA> 
#> 2 raw_basisOfRecord The basis of record as supplied by the data publisher http://rs.tdwg.org/dwc/terms/ver… fiel… <NA>
}\if{html}{\out{</div>}}

Search for all fields that have information for "marine"\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("marine") |> 
  head() # only show first 5 results
#> # A tibble: 6 × 4
#>   id      description                                           type   link                                            
#>   <chr>   <chr>                                                 <chr>  <chr>                                           
#> 1 biome   "A field indicating if the record has been located i… fields  <NA>                                           
#> 2 cl2105  "FAO Fishery Statistical Areas FAO Fishing areas (Ma… layers "http://www.fao.org/geonetwork/srv/en/main.home…
#> 3 cl10948 "CAPAD 2018 Terrestrial The Collaborative Australian… layers "http://www.environment.gov.au/fed/catalog/sear…
#> 4 cl10957 "CAPAD 2018 Marine The Collaborative Australian Prot… layers "http://www.environment.gov.au/fed/catalog/sear…
#> 5 el1056  "Endemism (Non-marine) Endemism (Non-marine) based o… layers ""                                              
#> 6 el957   "Averaged Topographic Relief  This data represents t… layers "http://www.ga.gov.au/meta/ANZCW0703015076.html"
}\if{html}{\out{</div>}}

Search for all Wordclim layers\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("worldclim", type = "layers")
#> # A tibble: 38 × 4
#>    id      description                                                                     type   link                 
#>    <chr>   <chr>                                                                           <chr>  <chr>                
#>  1 el10982 WorldClim 2.1: Temperature - warmest month max Max Temperature of Warmest Month layers https://www.worldcli…
#>  2 el10981 WorldClim 2.1: Temperature - seasonality Temperature Seasonality                layers https://www.worldcli…
#>  3 el10980 WorldClim 2.1: Temperature - isothermality Isothermality                        layers https://www.worldcli…
#>  4 el10990 WorldClim 2.1: Precipitation - wettest month Precipitation of Wettest Month     layers https://www.worldcli…
#>  5 el10995 WorldClim 2.1: Precipitation - warmest quarter Precipitation of Warmest Quarter layers https://www.worldcli…
#>  6 el10994 WorldClim 2.1: Precipitation - driest quarter Precipitation of Driest Quarter   layers https://www.worldcli…
#>  7 el10992 WorldClim 2.1: Precipitation - seasonality Precipitation Seasonality            layers https://www.worldcli…
#>  8 el10991 WorldClim 2.1: Precipitation - driest month Precipitation of Driest Month       layers https://www.worldcli…
#>  9 el10978 WorldClim 2.1: Temperature - annual mean Annual Mean Temperature                layers https://www.worldcli…
#> 10 el10983 WorldClim 2.1: Temperature - coldest month min Min Temperature of Coldest Month layers https://www.worldcli…
#> # … with 28 more rows
}\if{html}{\out{</div>}}



See a listing of all valid fields and layers\if{html}{\out{<div class="sourceCode r">}}\preformatted{show_all_fields()
#> # A tibble: 794 × 4
#>    id                    description                                                                        type  link 
#>    <chr>                 <chr>                                                                              <chr> <chr>
#>  1 abcdTypeStatus        ABCD field in use by herbaria                                                      fiel… <NA> 
#>  2 acceptedNameUsage     http://rs.tdwg.org/dwc/terms/acceptedNameUsage                                     fiel… <NA> 
#>  3 acceptedNameUsageID   http://rs.tdwg.org/dwc/terms/acceptedNameUsageID                                   fiel… <NA> 
#>  4 accessRights          <NA>                                                                               fiel… <NA> 
#>  5 assertionUserId       User ID of the person who has made an assertion about this record                  fiel… <NA> 
#>  6 assertions            A list of all assertions (user and system supplied) for a record resulting from d… fiel… <NA> 
#>  7 associatedMedia       http://rs.tdwg.org/dwc/terms/associatedMedia                                       fiel… <NA> 
#>  8 associatedOccurrences http://rs.tdwg.org/dwc/terms/associatedOccurrences                                 fiel… <NA> 
#>  9 associatedReferences  http://rs.tdwg.org/dwc/terms/associatedReferences                                  fiel… <NA> 
#> 10 associatedSequences   http://rs.tdwg.org/dwc/terms/associatedSequences                                   fiel… <NA> 
#> # … with 784 more rows
}\if{html}{\out{</div>}}

Use this function with \code{\link[=search_fields]{search_fields()}} to find fields and layers you wish
to use to filter your data queries
}

\references{
\itemize{
\item Darwin Core terms \url{https://dwc.tdwg.org/terms/}
\item ALA fields \url{https://api.ala.org.au/#ws72}
\item ALA assertions fields \url{https://api.ala.org.au/#ws81}
}
}
\seealso{
This function is used to pass valid arguments to
\code{\link[=galah_select]{galah_select()}} and \code{\link[=galah_filter]{galah_filter()}}.
To view valid values for a layer with categorical values, use
\code{\link[=find_field_values]{find_field_values()}}.
}
