% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_media.R
\name{show_all_media}
\alias{show_all_media}
\title{Show media in occurrence records}
\usage{
show_all_media(df)
}
\arguments{
\item{ids}{either a \code{tibble} from \code{atlas_occurrences} or a vector giving media ids (parsed or unparsed)}
}
\description{
Text here
}
