% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_assertions.R, R/show_all_atlases.R,
%   R/show_all_cached_files.R, R/show_all_collectory.R, R/show_all_fields.R,
%   R/show_all_minifunctions.R, R/show_all_profiles.R, R/show_all_ranks.R,
%   R/show_all_reasons.R, R/show_all_values.R
\name{show_all_assertions}
\alias{show_all_assertions}
\alias{show_all_atlases}
\alias{show_all_cached_files}
\alias{show_all_collections}
\alias{show_all_datasets}
\alias{show_all_providers}
\alias{show_all_fields}
\alias{show_all_minifunctions}
\alias{show_all_reasons}
\alias{show_all_ranks}
\alias{show_all_profiles}
\alias{show_all_values}
\title{Functions for showing detailed metadata}
\usage{
show_all_assertions()

show_all_atlases()

show_all_cached_files()

show_all_collections()

show_all_datasets()

show_all_providers()

show_all_fields()

show_all_profiles()

show_all_ranks()

show_all_reasons()

show_all_values(field)
}
\arguments{
\item{field}{\code{string}: field to return the categories for}
}
\value{
An object of class \code{tbl_df} and \code{data.frame} (aka a tibble)
}
\description{
Some text
}
\section{Examples}{


See all supported atlases\if{html}{\out{<div class="sourceCode r">}}\preformatted{show_all_atlases()
#> # A tibble: 6 x 3
#>   atlas     taxonomy_source taxonomy_info                                                    
#>   <chr>     <chr>           <chr>                                                            
#> 1 Australia ALA             https://bie.ala.org.au/                                          
#> 2 Austria   GBIF            https://www.gbif.org/dataset/d7dddbf4-2cf0-4f39-9b2a-bb099caae36c
#> 3 Guatemala GBIF            https://www.gbif.org/dataset/d7dddbf4-2cf0-4f39-9b2a-bb099caae36c
#> 4 Spain     GBIF            https://www.gbif.org/dataset/d7dddbf4-2cf0-4f39-9b2a-bb099caae36c
#> 5 Sweden    GBIF            https://www.gbif.org/dataset/d7dddbf4-2cf0-4f39-9b2a-bb099caae36c
#> 6 UK        NBN             https://www.nhm.ac.uk/our-science/data/uk-species.html
}\if{html}{\out{</div>}}

See a listing of all valid fields and layers\if{html}{\out{<div class="sourceCode r">}}\preformatted{show_all_fields()
#> # A tibble: 791 x 4
#>   id                  description        type   link 
#>   <chr>               <chr>              <chr>  <chr>
#> 1 acceptedNameUsage   Accepted name      fields <NA> 
#> 2 acceptedNameUsageID Accepted name      fields <NA> 
#> 3 accessRights        Access rights      fields <NA> 
#> 4 assertionUserId     Assertions by user fields <NA> 
#> # ... with 787 more rows
}\if{html}{\out{</div>}}

Show a listing of all accepted reasons for downloading occurrence data\if{html}{\out{<div class="sourceCode r">}}\preformatted{show_all_reasons()
#> # A tibble: 13 x 2
#>       id name                            
#>    <int> <chr>                           
#>  1     0 conservation management/planning
#>  2     1 biosecurity management/planning 
#>  3     2 environmental assessment        
#>  4     3 education                       
#>  5     4 scientific research             
#>  6     5 collection management           
#>  7     6 other                           
#>  8     7 ecological research             
#>  9     8 systematic research/taxonomy    
#> 10    10 testing                         
#> 11    11 citizen science                 
#> 12    12 restoration/remediation         
#> 13    13 species modelling
}\if{html}{\out{</div>}}

Add your download reason when configuring your session with \code{\link[=galah_config]{galah_config()}}\if{html}{\out{<div class="sourceCode r">}}\preformatted{galah_config(download_reason_id = 3)
}\if{html}{\out{</div>}}

Show a list of all available data quality profiles\if{html}{\out{<div class="sourceCode r">}}\preformatted{show_all_profiles()
#> # A tibble: 7 x 4
#>      id name                                                  shortName        description                          
#>   <int> <chr>                                                 <chr>            <chr>                                
#> 1     2 ALA General (pre-pipelines)                           ALA-legacy       "The default ALA profile filters out~
#> 2    35 Species Distribution Modelling (CSDM) (pre-pipelines) CSDM-legacy      "Base filters for the Collaborative ~
#> 3    44 Data licensed for all uses (pre-pipelines)            re-usable-legacy "Data licensed for re-use, including~
#> 4    92 ALA General                                           ALA              "The default ALA profile filters out~
#> 5   124 Species Distribution Modelling (CSDM)                 CSDM             "Base filters for the Collaborative ~
#> 6   133 Data licensed for all uses                            re-usable        "Data licensed for re-use, including~
#> 7   224 AVH                                                   AVH              "AVH data quality profile"
}\if{html}{\out{</div>}}

Values in the \code{shortName} column can be used with \code{\link[=galah_filter]{galah_filter()}} to
narrow your data query results\if{html}{\out{<div class="sourceCode r">}}\preformatted{galah_filter(profile == "ALA")
#> Warning: Invalid field(s) detected.
#> i See a listing of all valid fields with `show_all_fields()`.
#> i Search for the valid name of a desired field with `search_fields()`.
#> x Invalid field(s): profile.
#> # A tibble: 1 x 4
#>   variable logical value query              
#>   <chr>    <chr>   <chr> <chr>              
#> 1 profile  ==      ALA   "(profile:\\"ALA\\")"
}\if{html}{\out{</div>}}

Show a listing of all taxonomic ranks\if{html}{\out{<div class="sourceCode r">}}\preformatted{show_all_ranks()
#> # A tibble: 69 x 2
#>      id name        
#>   <int> <chr>       
#> 1     1 root        
#> 2     2 superkingdom
#> 3     3 kingdom     
#> 4     4 subkingdom  
#> # ... with 65 more rows
}\if{html}{\out{</div>}}

Use ranks with \code{\link[=galah_down_to]{galah_down_to()}} and \code{\link[=atlas_taxonomy]{atlas_taxonomy()}} to get taxonomic
trees\if{html}{\out{<div class="sourceCode r">}}\preformatted{galah_call() \%>\%
  galah_identify("fungi") \%>\%
  galah_down_to(subphylum) \%>\%
  atlas_taxonomy()
#>                        levelName
#> 1  Fungi                        
#> 2   ¦--Ascomycota               
#> 3   ¦   ¦--Pezizomycotina       
#> 4   ¦   ¦--Saccharomycotina     
#> 5   ¦   °--Taphrinomycotina     
#> 6   ¦--Basidiomycota            
#> 7   ¦   ¦--Agaricomycotina      
#> 8   ¦   ¦--Pucciniomycotina     
#> 9   ¦   ¦--Ustilaginomycotina   
#> 10  ¦   °--Wallemiomycotina     
#> 11  ¦--Mucoromycota             
#> 12  ¦   °--Glomeromycotina      
#> 13  °--Zygomycota               
#> 14      ¦--Entomophthoromycotina
#> 15      ¦--Kickxellomycotina    
#> 16      ¦--Mortierellomycotina  
#> 17      ¦--Mucoromycotina       
#> 18      °--Zoopagomycotina
}\if{html}{\out{</div>}}

Configure caching and create a query to cache with \code{\link[=galah_config]{galah_config()}}\if{html}{\out{<div class="sourceCode r">}}\preformatted{galah_config(caching = TRUE)
dat <- atlas_counts(group_by = galah_group_by(year))
}\if{html}{\out{</div>}}

Show a listing of previously cached files\if{html}{\out{<div class="sourceCode r">}}\preformatted{show_all_cached_files()
}\if{html}{\out{</div>}}
}

