% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_reasons.R
\name{show_all_reasons}
\alias{show_all_reasons}
\title{List valid download reasons}
\usage{
show_all_reasons()
}
\value{
An object of class \code{tbl_df} and \code{data.frame} (aka a tibble) of
valid download reasons, containing the id and name for each reason.
}
\description{
When downloading occurrence data with \code{\link[=atlas_occurrences]{atlas_occurrences()}} the
ALA APIs require a reason for download to be specified. By default, a
download reason of 'scientific research' is set for you, but if you wish to
change this you can do so with \code{\link[=galah_config]{galah_config()}}. Use this function
to view the list of download reason code and names. When specifying a reason,
you can use either the download code or name.
}
\seealso{
This function is helpful in setting up \code{\link[=galah_config]{galah_config()}}.
}
