% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise.data_request}
\alias{summarise.data_request}
\alias{summarize.data_request}
\title{Summarise for object of class \code{data_request}}
\usage{
\method{summarise}{data_request}(.data, type = c("record", "species"))

\method{summarize}{data_request}(.data, type = c("record", "species"))
}
\arguments{
\item{.data}{An object of class \code{data_request}, created using \code{\link[=galah_call]{galah_call()}}}

\item{type}{\code{string}: one of \code{c("record", "species")}. Defaults to
"record". If "species", the number of species matching the criteria will be
returned, if "record", the number of records matching the criteria will be
returned.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
