% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_counts.R
\name{ala_counts}
\alias{ala_counts}
\title{Counts for ALA records}
\usage{
ala_counts(taxa, filters, locations, group_by, limit = 100, caching = FALSE)
}
\arguments{
\item{taxa}{string or data.frame: one or more unique species identifiers.
Use \code{\link{ala_taxa}} to generate a data.frame for this argument.}

\item{filters}{data.frame: generated by a call to \code{\link{select_filters}}}

\item{locations}{string: generated by a call to \code{\link{select_locations}}}

\item{group_by}{field to count by}

\item{limit}{numeric: maximum number of categories to return. 100 by default.
If limit is NULL, all results are returned. For some categories this will
take a while.}

\item{caching}{logical: should the results be cached/use cache? Only used if
`group_by` is supplied.}
}
\value{
either single integer of total counts, or a data.frame of counts by
`group_by` field, if specified.
}
\description{
`ala_counts` takes filters in the same format as\code{\link{ala_occurrences}}.
Provides either a total count of records matching the criteria, or a
data.frame of counts matching the criteria, broken down by the `group_by` field.
}
\examples{
# Count of total records in the ALA
ala_counts()
# Counts by state and territory
ala_counts(group_by = "state")
}
