% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_counts.R
\name{ala_counts}
\alias{ala_counts}
\title{Summary counts}
\usage{
ala_counts(taxa, filters, locations, group_by, limit = 100, caching = FALSE)
}
\arguments{
\item{taxa}{data.frame: generated by a call to \code{\link{ala_taxa}}. This
argument also accepts a vector of unique species identifiers.}

\item{filters}{data.frame: generated by a call to \code{\link{select_filters}}}

\item{locations}{string: generated by a call to \code{\link{select_locations}}}

\item{group_by}{field to count by}

\item{limit}{numeric: maximum number of categories to return. 100 by default.
If limit is NULL, all results are returned. For some categories this will
take a while.}

\item{caching}{logical: should the results be cached/use cache? Only used if
`group_by` is supplied.}
}
\value{
\itemize{
 \item{A single count, if \code{group_by} is not specified or,}
 \item{A data.frame of counts by \code{group_by} field, if it is specified}
}
}
\description{
\code{ala_counts} takes arguments in the same format as
\code{\link{ala_occurrences}}.
Provides either a total count of records matching the criteria, or a
data.frame of counts matching the criteria, broken down by the `group_by`
field.
}
\examples{
# With no arguments, return the total number of records in the ALA
ala_counts()

# Group counts by state and territory
ala_counts(group_by = "stateProvince")

# Count records matching a filter
ala_counts(filters = select_filters(basisOfRecord = "FossilSpecimen"))

}
