% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_media.R
\name{ala_media}
\alias{ala_media}
\title{Image and sound downlaod}
\usage{
ala_media(
  identifier,
  identifier_type = "media",
  download_dir,
  media_type = c("image", "sound")
)
}
\arguments{
\item{identifier}{string: a single or vector of occurrence or media
identifiers. The type is specified by \code{identifier_type}}

\item{identifier_type}{string: one of \code{c("occurrence", "media")}.
Defaults to "media".}

\item{download_dir}{string: path to directory to store the downloaded media
in}

\item{media_type}{string: type of media to download, one or both of 
\code{c("image", "sound")}`. Defaults to both.}
}
\value{
dataframe of media information
}
\description{
\code{\link{ala_occurrences}} returns a \code{recordID} field- passing this
to \code{ala_media} will return images and sounds for all the records.
}
\examples{
\dontrun{
# Occurrences of galahs with sound recordings
occ <- ala_occurrences(taxa = ala_taxa("Phascolarctos cinereus"),
                       filters = select_filters(multimedia = "Sound"))
sounds <- ala_media(occ$recordID, identifier_type = "occurrence",
                    download_dir = tempdir())
}
}
