% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_media.R
\name{ala_media}
\alias{ala_media}
\title{Images, sounds and videos}
\usage{
ala_media(
  taxa = NULL,
  filter = NULL,
  location = NULL,
  select = galah_select(group = "basic"),
  columns = NULL,
  download_dir,
  refresh_cache = FALSE
)
}
\arguments{
\item{taxa}{\code{data.frame}: generated by a call to
\code{\link{select_taxa}()}. This argument also accepts a vector of unique
species identifiers.}

\item{filter}{\code{data.frame}: generated by a call to
\code{\link{galah_filter}()}}

\item{location}{\code{string}: generated by a call to
\code{\link{galah_location}()}}

\item{select}{\code{data.frame}: generated by a call to
\code{\link{galah_select}()}}

\item{columns}{\code{data.frame}: DEPRECATED - use \code{select} instead}

\item{download_dir}{\code{string}: path to directory to store the downloaded
media in}

\item{refresh_cache}{\code{logical}: if set to `TRUE` and 
`galah_config(caching = TRUE)` then files cached from a previous query will 
be replaced by the current query}
}
\value{
\code{data.frame} of metadata of the downloaded media
}
\description{
In addition to text data describing individual occurrences and their attributes,
ALA stores images, sounds and videos associated with a given record. \code{ala_media}
allows download of any and all of the media types.
}
\details{
\code{\link{ala_occurrences}()} works by first finding all occurrence records
matching the filter which contain media, then downloading the metadata for the
media and the media files. \code{\link{galah_filter}()} can take both filter
relating to occurrences (e.g. basis of records), and filter relating to media
(e.g. type of licence).
It may be beneficial when requesting a large number of records to show a progress
bar by setting \code{verbose = TRUE} in \code{\link{galah_config}()}.
}
\examples{
\dontrun{
# Download Regent Honeyeater multimedia
media_data <- ala_media(
    taxa = select_taxa("Regent Honeyeater"),
    filter = galah_filter(year = 2011),
    download_dir = "media")

# Specify a single media type to download
media_data <- ala_media(
     taxa = select_taxa("Eolophus Roseicapilla"),
     filter = galah_filter(multimedia = "Sound"))

# Filter to only records with a particular licence type
media_data <- ala_media(
      taxa = select_taxa("Ornithorhynchus anatinus"),
      filter = galah_filter(year = 2020,
      license = "http://creativecommons.org/licenses/by-nc/4.0/")
)
# Check how many records have media files
ala_counts(
     filter = galah_filter(multimedia = c("Image","Sound","Video")),
     group_by = "multimedia"
)
}
}
\seealso{
\code{\link{ala_counts}} to find the number of records with media- note this
is not necessarily the same as the number of media files, as each record can have
more than one media file associated with it (see examples section for how to do this).
}
