% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_occurrences.R
\name{ala_occurrences}
\alias{ala_occurrences}
\title{Occurrence records}
\usage{
ala_occurrences(taxa, filters, locations, columns, mint_doi = FALSE)
}
\arguments{
\item{taxa}{data.frame: generated by a call to \code{\link{select_taxa}}. This
argument also accepts a vector of unique species identifiers.}

\item{filters}{data.frame: generated by a call to \code{\link{select_filters}}}

\item{locations}{string: generated by a call to \code{\link{select_locations}}}

\item{columns}{data.frame: generated by a call to \code{\link{select_columns}}}

\item{mint_doi}{logical: by default no DOI will be generated. Set to
true if you intend to use the data in a publication or similar}
}
\value{
data.frame of occurrences
}
\description{
Large downloads can take a long time to process. To show a progress bar
for the download, set \code{verbose = TRUE} in \code{\link{ala_config}}. The
maximum number of records that can be downloaded in one go is 50 million.
}
\examples{
\dontrun{
# Search for occurrences matching a taxon identifier
occ <- ala_occurrences(taxa = select_taxa("Reptilia"))

# Search for occurrences in a year range
occ <- ala_occurrences(filters = select_filters(year = seq(2010, 2020)))

# Search for occurrences in a WKT-specified area
polygon <- "POLYGON((146.24960 -34.05930,146.37045 -34.05930,146.37045 -34.152549,146.24960 -34.15254,146.24960 -34.05930))"
occ <- ala_occurrences(locations = select_locations(wkt = polygon))
}
}
