% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_taxa.R
\name{select_taxa}
\alias{select_taxa}
\title{Taxon information}
\usage{
select_taxa(
  term,
  term_type = "name",
  return_children = FALSE,
  include_counts = FALSE,
  caching = FALSE
)
}
\arguments{
\item{term}{string: named list or dataframe: search term(s)}

\item{term_type}{string: specifies which type of terms are provided in
`term`. One of name \code{`c('name', 'identifier')`}. Default behaviour is
to search by `name`. `identifier` refers to the unique identifier for a
taxon.}

\item{return_children}{logical: return child concepts for the provided
term(s)?}

\item{include_counts}{logical: return occurrence counts for all taxa
found? \code{FALSE} by default}

\item{caching}{logical: should the results be cached? \code{FALSE} by
default}
}
\value{
data.frame of taxon information
}
\description{
The resulting dataframe of taxon information can be passed directly to
\code{\link{ala_occurrences}} to filter records to the specified taxa.
}
\examples{
\dontrun{
# Search using a single term
select_taxa(term = "Reptilia")

# Search with multiple ranks. This is required if a single term is a homonym.
select_taxa(list(kingdom = "Plantae", genus = "Microseris"),
         return_children = TRUE, include_counts = TRUE)

# Search using an unique taxon identifier
select_taxa(term = "https://id.biodiversity.org.au/node/apni/2914510",
         term_type = "identifier")
}
}
