% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_query.R
\name{as_query.data_request}
\alias{as_query.data_request}
\alias{as_query}
\alias{as_query.list}
\alias{as_query.metadata_request}
\alias{as_query.files_request}
\title{Convert an object to class \code{query}}
\usage{
as_query(x, ...)

\method{as_query}{list}(x)

\method{as_query}{data_request}(x, mint_doi = FALSE, ...)

\method{as_query}{metadata_request}(x, ...)

\method{as_query}{files_request}(x, thumbnail = FALSE, ...)
}
\arguments{
\item{x}{An object to convert to a \code{query}. Supported classes are the same
as those produced by \code{\link[=galah_call]{galah_call()}}, namely \code{data_request},
\code{metadata_request} or \code{files_request}.}

\item{...}{Other arguments, currently ignored}

\item{mint_doi}{Logical: should a DOI be minted for this download? Only
applies to \code{type = "occurrences"} when atlas chosen is "ALA".}

\item{thumbnail}{Logical: should thumbnail-size images be returned? Defaults
to \code{FALSE}, indicating full-size images are required.}
}
\value{
An object of class \code{query}, which is a list-like object containing
two or more of the following slots:
\itemize{
\item \code{type}: The type of query, serves as a lookup to the corresponding field in \code{show_all(apis)}
\item \code{url}: Either:
\itemize{
\item a length-1 character giving the API to be queried; or
\item a \code{tibble()} containing at least the field \code{url} and optionally others
}
\item \code{headers}: headers to be sent with the API call
\item \code{body}: body section of the API call
\item \code{options}: options section of the API call
\item Any other information retained from the preceeding \verb{_request} object (see \code{\link[=galah_call]{galah_call()}})
}
}
\description{
Functionally similar to \code{\link[=collapse.data_request]{collapse()}}, but
without passing through \code{\link[=coalesce]{coalesce()}} first. Primarily an internal function,
but exported for clarity and debugging purposes.
}
\details{
Typically, queries in galah are piped using \code{\link[=galah_call]{galah_call()}}, which builds
an object of class \code{"data_request"}, \code{"metadata_request"} or
\code{"files_request"}. All these objects can be converted to class \code{"query"}
using \code{\link[=collapse.data_request]{collapse()}}. However,
\code{\link[=collapse.data_request]{collapse()}} first calls
\code{\link[=coalesce.data_request]{coalesce()}}, which expands to an
object of class \code{"query_set"} \emph{before} evaluating
\code{\link[=collapse.data_request]{collapse()}}. In this context,
\code{\link[=as_query]{as_query()}} serves two purposes: externally, it can be called to convert
directly to class \code{"query"} without running checks; and internally it allows
a query to be appended to a \code{"query_set"} without calling causing an
infinite loop.

For simple cases, this gives the same result as running
\code{\link[=collapse.data_request]{collapse()}} while the \code{run_checks}
argument of \code{\link[=galah_config]{galah_config()}} is set to \code{FALSE}, but is slightly faster. For
complex cases, however, it is likely to generate irresolvable API calls,
because e.g. taxonomic queries are not parsed before the URL is built. It
should therefore be used with care.
}
\seealso{
To open a piped query, see \code{\link[=galah_call]{galah_call()}}. For alternative
operations on \verb{_request} objects, see \code{\link[=coalesce]{coalesce()}},
\code{\link[=collapse.data_request]{collapse()}},
\code{\link[=compute.data_request]{compute()}} or
\code{\link[=collect.data_request]{collect()}}.
}
