% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_citation.R
\name{atlas_citation}
\alias{atlas_citation}
\title{Generate a citation for occurrence data}
\usage{
atlas_citation(data)
}
\arguments{
\item{data}{data.frame: occurrence data generated by
\code{\link[=atlas_occurrences]{atlas_occurrences()}}}
}
\value{
A string containing the citation for that dataset.
}
\description{
If a \code{data.frame} was generated using \code{\link[=atlas_occurrences]{atlas_occurrences()}},
and the \code{mint_doi} argument was set to \code{TRUE}, the DOI associated
with that dataset is appended to the resulting \code{data.frame} as an
attribute. This function simply formats that DOI as a citation that can be
included in a scientific publication. Please also consider citing this
package, using the information in \code{citation("galah")}.
}
