% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_media.R
\name{atlas_media}
\alias{atlas_media}
\title{Get metadata on images, sounds and videos}
\usage{
atlas_media(
  request = NULL,
  identify = NULL,
  filter = NULL,
  select = NULL,
  geolocate = NULL,
  data_profile = NULL
)
}
\arguments{
\item{request}{optional \code{data_request} object: generated by a call to
\code{\link[=galah_call]{galah_call()}}.}

\item{identify}{\code{data.frame}: generated by a call to
\code{\link[=galah_identify]{galah_identify()}}.}

\item{filter}{\code{data.frame}: generated by a call to
\code{\link[=galah_filter]{galah_filter()}}}

\item{select}{\code{list}: generated by a call to \code{\link[=galah_select]{galah_select()}}}

\item{geolocate}{\code{string}: generated by a call to
\code{\link[=galah_geolocate]{galah_geolocate()}}}

\item{data_profile}{\code{string}: generated by a call to
\code{\link[=galah_apply_profile]{galah_apply_profile()}}}
}
\value{
An object of class \code{tbl_df} and \code{data.frame} (aka a tibble)
of metadata of the requested media.
}
\description{
In addition to text data describing individual occurrences and their
attributes, ALA stores images, sounds and videos associated with a given
record. \code{atlas_media} displays metadata for any and all of the media types.
}
\details{
\code{\link[=atlas_media]{atlas_media()}} works by first finding all occurrence records
matching the filter which contain media, then downloading the metadata for the
media. To actually download the files themselves, use \code{\link[=collect_media]{collect_media()}}.
It may be beneficial when requesting a large number of records to show a progress
bar by setting \code{verbose = TRUE} in \code{\link[=galah_config]{galah_config()}}.
}
\examples{
\dontrun{
# Download Regent Honeyeater records with multimedia attached
galah_call() |>
  galah_identify("Regent Honeyeater") |>
  galah_filter(year == 2011) |>
  atlas_media()

# Download multimedia
galah_call() |>
  galah_identify("Regent Honeyeater") |>
  galah_filter(year == 2011) |>
  atlas_media() |>
  collect_media(path = "folder/your-directory")

# Specify a single media type to download
galah_call() |>
  galah_identify("Eolophus Roseicapilla") |>
  galah_filter(multimedia == "Sound") |>
  atlas_media()

# It's good to check how many records have media files before downloading
galah_call() |>
  galah_filter(multimedia == c("Image", "Sound", "Video")) |>
  galah_group_by(multimedia) |>
  atlas_counts()
  
  
# post version 2.0, it is possible to run all steps in sequence
# first, get occurrences, making sure to include media fields:
occurrences_df <- request_data() |>
  identify("Regent Honeyeater") |>
  filter(!is.na(images), year == 2011) |>
  select(group = "media") |>
  collect()
 
# second, get media metadata
media_info <- request_metadata() |>
  filter(media == occurrences_df) |>
  collect()
  
# the two steps above + `right_join()` are synonmous with `atlas_media()`
# third, get images
request_files() |>
  filter(media == media_df) |>
  collect(thumbnail = TRUE)
  
# step three is synonymous with `collect_media()`
}
}
\seealso{
\code{\link[=atlas_counts]{atlas_counts()}} to find the number of records with media; but note this
is not necessarily the same as the number of media files, as each record can have
more than one media file associated with it (see examples section for how to do this).
}
