% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_species.R
\name{atlas_species}
\alias{atlas_species}
\title{Return species lists}
\usage{
atlas_species(
  taxa = NULL,
  filter = NULL,
  location = NULL,
  refresh_cache = FALSE
)
}
\arguments{
\item{taxa}{\code{data.frame}: generated by a call to
\code{\link{select_taxa}()}. This argument also accepts a vector of unique
species identifiers.}

\item{filter}{\code{data.frame}: generated by a call to
\code{\link{galah_filter}()}}

\item{location}{\code{string}: generated by a call to
\code{\link{galah_location}()}}

\item{refresh_cache}{\code{logical}: if set to `TRUE` and 
`galah_config(caching = TRUE)` then files cached from a previous query will 
be replaced by the current query}
}
\value{
A \code{data.frame} of matching species. The \code{data.frame} object 
has attributes listing of the user-supplied arguments of the \code{data_request} 
(i.e., taxa, filter, location, columns)
}
\description{
While there are reasons why users may need to check every record meeting their
search criteria (i.e. using \code{\link{atlas_occurrences}}), a common use case
is to simply identify which species occur in a specified region, time period,
or taxonomic group. This function returns a \code{data.frame} with one row
per species, and columns giving associated taxonomic information.
}
\details{
The primary use case of this function is to extract species-level information
given a set of criteria defined by \code{\link{select_taxa}()},
\code{\link{galah_filter}()} or \code{\link{galah_location}()}. If the purpose
is simply to get taxonomic information that is not restricted by filtering,
then \code{\link{select_taxa}()} is more efficient. Similarly, if counts are
required that include filter but without returning taxonomic detail, then
\code{\link{atlas_counts}()} is more efficient (see examples).
}
\examples{
\dontrun{

# Look up genus "Heleioporus" in the ALA
select_taxa("Heleioporus")

# Find how many records there are for this genus
atlas_counts(select_taxa("Heleioporus"))

# Get taxonomic information on all species within this genus
# (every row is a species with associated taxonomic data)
atlas_species(select_taxa("Heleioporus"))
}
}
