% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlas_taxonomy.R
\name{atlas_taxonomy}
\alias{atlas_taxonomy}
\alias{atlas_taxonomy.data_request}
\alias{atlas_taxonomy.default}
\title{Search taxonomic trees}
\usage{
atlas_taxonomy(...)

\method{atlas_taxonomy}{data_request}(request, ...)

\method{atlas_taxonomy}{default}(taxa, down_to)
}
\arguments{
\item{taxa}{The identity of the clade for which a taxonomic
hierarchy should be returned. Should be specified using an object of class
\code{data.frame} and \code{ala_id}, as returned from
\code{\link[=search_taxa]{search_taxa()}}.}

\item{down_to}{The identity of the clade at which the downwards search
should stop. Should be specified using an object of class
\code{character} and \code{galah_down_to}, as returned from
\code{\link[=galah_down_to]{galah_down_to()}}. Also accepts a string.}
}
\value{
A tree consisting of objects of class \code{Node}, containing the
requested taxonomy. Each node contains the following attributes:
\itemize{
\item \code{name}: The scientific name of the taxon in question
\item \code{rank}: The taxonomic rank to which that taxon belongs
\item \code{guid}: A unique identifier used by the ALA
\item \code{authority}: The source of the taxonomic name & identifier
}
}
\description{
The ALA has its' own internal taxonomy that is derived from authoritative
sources. \code{atlas_taxonomy} provides a means to query
that taxonomy, returning a tree (class \code{Node}) showing which lower
clades are contained within the specified taxon.
}
\details{
The approach used by this function is recursive, meaning that it
becomes slow for large queries such as
\code{atlas_taxonomy(search_taxa("Plantae"), down_to = galah_down_to(species))}.
Although the inputs to \code{search_taxa} and \code{down_to} are
case-insensitive, node names are always returned in title case.
}
\examples{
atlas_taxonomy(search_taxa("chordata"), down_to = galah_down_to(class))

}
\seealso{
\code{\link[=search_taxa]{search_taxa()}} to search for an individual taxon;
\code{\link[=show_all_ranks]{show_all_ranks()}} for valid ranks used to specify the \code{down_to}
argument.
}
