% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Set up authentication}
\usage{
authenticate(
  .data,
  email = NULL,
  username = NULL,
  password = NULL,
  use_jwt = FALSE,
  cache_jwt = FALSE,
  ...
)
}
\arguments{
\item{.data}{An object of class \code{data_request} or \code{metadata_request}}

\item{email}{(string) Email address registered with the selected organisation}

\item{username}{(string) Registered username (GBIF only)}

\item{password}{(string) Registered password (GBIF only)}

\item{use_jwt}{(logical) Should an OAuth workflow be used for authentication?
Only supported for Flemish and Australian atlases. Defaults to \code{FALSE}}

\item{cache_jwt}{(logical) Should JWT tokens be cached to disk? Defaults
to \code{FALSE}}

\item{...}{Other arguments, currently ignored}
}
\value{
An object of the same class as supplied, but with an added
\code{authenticate} slot.
}
\description{
Authenticate a request, either by sending a registered email address
(and, for GBIF, password and username); or by loggin in via the
browser to generate a JWT token. Note that while handling this manually
in-pipe is the most transparent approach, this function is also used to
pass cached information stored via \code{\link[=galah_config]{galah_config()}} to a query.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}.
}
\examples{
\dontrun{
# use `galah_config()` to set for all occurrence queries
galah_config(authenticate = TRUE)

x <- galah_call() |>
  identify("Wollemia nobilis") |>
  collect()

# use in-pipe for more control
x <- galah_call() |>
  identify("Wollemia nobilis") |>
  authenticate() |>
  collect()
}
}
