% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\alias{use_authentication}
\title{Set up authentication}
\usage{
authenticate()

use_authentication(.data, cache_disk = FALSE)
}
\description{
This is early-stage code. It's purpose is to trigger the browser to generate
a JWT token. \code{\link[=authenticate]{authenticate()}} sets \code{galah_config(authenticate = TRUE)}, but
differs in triggering authentication on execution (which \code{galah_config()}
does not). \code{\link[=use_authentication]{use_authentication()}} is an in-pipe method for setting
authentication and associated behaviour. It is primarily intended for
internal use, but is exported for completeness and debugging purposes.
}
\details{
By default, authentication is only triggered during occurrence downloads
and query uploads. This can be overrided by calling \code{\link[=use_authentication]{use_authentication()}}.
}
