% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_cached_files.R
\name{clear_cached_files}
\alias{clear_cached_files}
\title{Clear previously cached files}
\usage{
clear_cached_files()
}
\description{
Deletes cached files within the cached file directory and their query 
metadata
}
\examples{
\dontrun{
## configure caching and create a query to cache
galah_config(caching = TRUE)
dat <- ala_counts(group_by = "year")

## clear cached files directory
clear_cached_files()
}
}
