% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_cached_files.R
\name{clear_cached_files}
\alias{clear_cached_files}
\title{Clear previously cached files}
\usage{
clear_cached_files()
}
\value{
No return value; called for side effect of removing files
}
\description{
Deletes cached files within the cached file directory and their query
metadata
}
\examples{
\dontrun{
# First set caching to true with [galah_config()]
galah_config(caching = TRUE)

# Then create a data query.
# The data you download will be cached in a temporary directory.
dat <- atlas_counts(group_by = galah_group_by(year))

# To clear your cached files directory, use `clear_cached_files()`
clear_cached_files()
}

}
