% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse.data_request}
\alias{collapse.data_request}
\title{collapse for galah}
\usage{
\method{collapse}{data_request}(.data, type = c("counts", "species", "occurrences", "media"))
}
\arguments{
\item{.data}{An object of class \code{data_request}.}
}
\description{
The \code{collapse()} function is ported from \code{dplyr}, and is used to convert
a piped query into an object that can be evaluated by \code{url_GET()} and related
methods. This is used internally for constructing actionable queries, but
is exported here for debugging purposes.
}
