% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect.data_request}
\alias{collect.data_request}
\alias{collect.data_response}
\title{Collect data from the selected atlas}
\usage{
\method{collect}{data_request}(.data, what)

\method{collect}{data_response}(.data, wait = FALSE)
}
\arguments{
\item{.data}{An object of class \code{data_response}, created using
\code{\link[=compute.data_request]{compute.data_request()}}}

\item{what}{string: what kind of data should be returned. Must be one of
\code{"counts"}, \code{"species"}, \code{"occurrences"} or \code{"media"}.}
}
\value{
A \code{tibble} containing requested data
}
\description{
For \code{type = "media"} or \code{"occurrences"}, object must first have been passed
to the specified atlas using \code{compute()}. For \code{type = "counts"} this is not
required; in fact \code{collect()} is synonymous with \code{count()} in this instance.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}.
}
\details{
This function has three extensions in \code{galah}. \code{collect.data_request()} is
designed to be called at the end of a pipe, \code{collect.data_response()} works
after calls to \code{compute}, and \code{collect.data_query()} works after calls to
\code{collapse()}.
}
\seealso{
\code{\link[=atlas_occurrences]{atlas_occurrences()}}
}
