% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R, R/collect.R, R/compute.R
\name{collapse.data_request}
\alias{collapse.data_request}
\alias{collapse.metadata_request}
\alias{collapse.files_request}
\alias{collect.data_request}
\alias{collect.data_query}
\alias{collect.data_response}
\alias{collect.metadata_request}
\alias{collect.metadata_query}
\alias{collect.metadata_response}
\alias{collect.files_request}
\alias{collect.files_query}
\alias{collect.files_response}
\alias{compute.data_request}
\alias{compute.data_query}
\alias{compute.metadata_request}
\alias{compute.metadata_query}
\alias{compute.files_request}
\alias{compute.files_query}
\title{Force computation of a database query}
\usage{
\method{collapse}{data_request}(.data)

\method{collapse}{metadata_request}(.data)

\method{collapse}{files_request}(.data, thumbnail = FALSE)

\method{collect}{data_request}(.data, wait = FALSE, file = NULL)

\method{collect}{data_query}(.data, wait = TRUE, file = NULL)

\method{collect}{data_response}(.data, wait = TRUE, file = NULL)

\method{collect}{metadata_request}(.data)

\method{collect}{metadata_query}(.data)

\method{collect}{metadata_response}(.data)

\method{collect}{files_request}(.data)

\method{collect}{files_query}(.data)

\method{collect}{files_response}(.data, file = NULL)

\method{compute}{data_request}(.data)

\method{compute}{data_query}(.data)

\method{compute}{metadata_request}(.data)

\method{compute}{metadata_query}(.data)

\method{compute}{files_request}(.data)

\method{compute}{files_query}(.data)
}
\arguments{
\item{.data}{An object of class \code{data_request}, \code{data_query} or
\code{data_response}}

\item{wait}{logical; should \code{galah} ping the selected url until computation
is complete? Defaults to \code{FALSE}.}

\item{file}{(optional) file name. If not given will be \code{data} with date and
time added. File path is always given by \code{galah_config()$package$directory}.}
}
\value{
\code{collect()} returns a \code{tibble} containing requested data. \code{compute()}
returns an object of class \code{data_response}. \code{collapse()} returns an object of
class \code{data_query}.
}
\description{
The \code{collapse()}, \code{compute()} and \code{collect()} functions are borrowed from
\code{dplyr}, and underpin every \code{atlas_} function in \code{galah}.

\code{collapse()} constructs a valid query so it can be
inspected before being sent. \code{compute()} sends the query so that it can
be calculated server-side in the specified atlas. \code{collect()} returns the
resulting \code{tibble} once it is complete.
}
\details{
The \code{collect()} function has three extensions in \code{galah}:
\itemize{
\item \code{collect.data_request()} is designed to be called at the end of a pipe
\item \code{collect.data_response()} is called after \code{compute()}
\item \code{collect.data_query()} is called after \code{collapse()}.
}
}
