% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R, R/collect.R, R/compute.R
\name{collapse.data_request}
\alias{collapse.data_request}
\alias{collapse.metadata_request}
\alias{collapse.files_request}
\alias{collect.data_request}
\alias{collect.data_query}
\alias{collect.data_response}
\alias{collect.metadata_request}
\alias{collect.metadata_query}
\alias{compute.data_request}
\alias{compute.data_query}
\title{Force computation of a database query}
\usage{
\method{collapse}{data_request}(.data)

\method{collapse}{metadata_request}(.data)

\method{collapse}{files_request}(.data)

\method{collect}{data_request}(.data)

\method{collect}{data_query}(.data)

\method{collect}{data_response}(.data, wait)

\method{collect}{metadata_request}(.data)

\method{collect}{metadata_query}(.data)

\method{compute}{data_request}(.data)

\method{compute}{data_query}(.data)
}
\arguments{
\item{.data}{An object of class \code{data_request}, \code{data_query} or
\code{data_response}}

\item{wait}{logical; should \code{galah} ping the selected url until computation
is complete? Defaults to \code{FALSE}.}
}
\value{
\code{collect()} returns a \code{tibble} containing requested data. \code{compute()}
returns an object of class \code{data_response}. \code{collapse()} returns an object of
class \code{data_query}.
}
\description{
These functions are borrowed from \code{dplyr}, and underpin every \code{atlas_}
function in \code{galah}. \code{collapse()} calculates a valid query so it can be
inspected before being sent. \code{compute()} sends the query so that it can
be calculated server-side in the specified atlas. \code{collect()} returns the
resulting \code{tibble} once it is complete.
}
\details{
This function has three extensions in \code{galah}. \code{collect.data_request()} is
designed to be called at the end of a pipe, \code{collect.data_response()} works
after calls to \code{compute}, and \code{collect.data_query()} works after calls to
\code{collapse()}.
}
