% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_media.R
\name{collect_media}
\alias{collect_media}
\title{Collect media files}
\usage{
collect_media(df, thumbnail = FALSE)
}
\arguments{
\item{df}{tibble returned by \code{atlas_media()} or a pipe starting with
\code{request_data(type = "media")}}

\item{thumbnail}{logical: should the download return thumbnails (TRUE) or
full size images (FALSE, the default)}
}
\value{
Available image & media files downloaded to a user local directory.
}
\description{
This function downloads full-sized or thumbnail images and media files using
information from \code{atlas_media} to a local directory.
}
\examples{
\dontrun{
# Use `atlas_media()` to return a `tibble` of records that contain media
x <- galah_call() |> 
  galah_identify("perameles") |>
  galah_filter(year == 2015) |>
  atlas_media()

# To download media files, add `collect_media()` to the end of a query 
galah_config(directory = "media_files")
collect_media(x)
}
}
