% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_occurrences.R
\name{collect_occurrences}
\alias{collect_occurrences}
\title{Collect occurrence records from a pre-existing DOI or URL}
\usage{
collect_occurrences(url, doi)
}
\arguments{
\item{url}{\code{string}: Retrieve occurrence
records previously downloaded from the ALA, using the URL provided via email.}

\item{doi}{\code{string}: Retrieve occurrence
records previously downloaded from the ALA, using the DOI generated by the
data.}
}
\value{
An object of class \code{tbl_df} and \code{data.frame} (aka a tibble) of
occurrences
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Download occurrence records using an existing DOI or URL. Pre-existing DOIs
and URLs come from previously generated downloads using \code{atlas_occurrences}
or online.
}
