% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound.R
\name{compound}
\alias{compound}
\alias{compound.data_request}
\alias{compound.metadata_request}
\alias{compound.files_request}
\alias{compound.prequery}
\alias{compound.query}
\alias{compound.query_set}
\title{Force evaluation of a database query}
\usage{
compound(x, ...)

\method{compound}{data_request}(x, mint_doi = FALSE, ...)

\method{compound}{metadata_request}(x, ...)

\method{compound}{files_request}(x, ...)

\method{compound}{prequery}(x, mint_doi = FALSE, ...)

\method{compound}{query}(x, ...)

\method{compound}{query_set}(x, ...)
}
\arguments{
\item{x}{An object to be compounded. Works for \code{data_request},
\code{metadata_request}, \code{file_request}, \code{query} or \code{prequery}.}

\item{...}{Other arguments passed to \code{\link[=capture]{capture()}}.}

\item{mint_doi}{Logical: should a DOI be minted for this download? Only
applies to \code{type = "occurrences"}, and only for supported atlases.}
}
\value{
An object of class \code{query_set}, which is simply a list of all \code{query}
objects required to properly evaluate the specified request. Objects are
listed in the order in which they will be evaluated, meaning the query
that the user has actually requested will be placed last.
}
\description{
\code{\link[=compound]{compound()}} is an S3 generic function intended to be called before
\code{\link[=collapse]{collapse()}}. It is important as it shows the full set of queries
required to properly evaluate the user's request. This is often broader
than the single query returned by \code{\link[=collapse]{collapse()}}. If, for example,
the user's query includes a call to
\code{\link[=identify.data_request]{identify()}}, then a taxonomic query
is required to run \emph{before} the 'final' query is attempted. In relation to
other functions that manipulate \verb{_request} objects, \code{\link[=compound]{compound()}} is called
within \code{\link[=collapse.data_request]{collapse()}}, and itself
calls \code{\link[=capture]{capture()}} internally where required.
}
\seealso{
To open a piped query, see \code{\link[=galah_call]{galah_call()}}. For alternative
operations on \verb{_request} objects, see \code{\link[=capture]{capture()}},
\code{\link[=collapse.data_request]{collapse()}},
\code{\link[=compute.data_request]{compute()}} or
\code{\link[=collect.data_request]{collect()}}.
}
