% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-compute.R
\name{compute.data_request}
\alias{compute.data_request}
\alias{compute.metadata_request}
\alias{compute.files_request}
\alias{compute.query}
\alias{compute.query_set}
\title{Compute a query}
\usage{
\method{compute}{data_request}(x, ...)

\method{compute}{metadata_request}(x, ...)

\method{compute}{files_request}(x, ...)

\method{compute}{query}(x, ...)

\method{compute}{query_set}(x, ...)
}
\arguments{
\item{x}{An object of class \code{data_request}, \code{metadata_request} or
\code{files_request} (i.e. constructed using a pipe) or \code{query}
(i.e. constructed by \code{collapse()})}

\item{...}{Arguments passed on to other methods}
}
\value{
An object of class \code{computed_query}, which is identical to class
\code{query} except for occurrence data, where it also contains information on the
status of the request.
}
\description{
This function sends a request for information to a server. This is only
useful for processes that run a server-side process, as it separates the
submission of the request from its' retrieval. Within galah, this is used
exclusively for generating occurrence queries, where calling
\code{\link[=compute.data_request]{compute()}} and then passing
the resulting \code{query} object to \code{\link[=collect.data_request]{collect()}}
at a later time can be preferable to calling \code{\link[=atlas_occurrences]{atlas_occurrences()}}, which
prevents execution of new code until the server-side process is complete.
}
\seealso{
To open a piped query, see \code{\link[=galah_call]{galah_call()}}. For alternative
operations on \verb{_request} objects, see \code{\link[=as_query]{as_query()}}, \code{\link[=coalesce]{coalesce()}},
\code{\link[=collapse.data_request]{collapse()}},
\code{\link[=collect.data_request]{collect()}}.
}
