% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R, R/print.R
\name{compute.data_request}
\alias{compute.data_request}
\alias{compute.data_query}
\alias{print.data_response}
\title{Join the queue to download records from the chosen atlas}
\usage{
\method{compute}{data_request}(.data, type)

\method{compute}{data_query}(.data)

\method{print}{data_response}(x, ...)
}
\arguments{
\item{.data}{An object of class \code{data_request} created using
\code{\link[=galah_call]{galah_call()}}, or of class \code{data_query}, created using \code{collapse.data_request()}}

\item{x}{an object of class \code{data_response}}
}
\value{
An object of class \code{data_response}
}
\description{
There are two versions of this function. If \code{compute} is called at the end of
a pipe started with \code{galah_call()}, then \code{compute.data_request()} is used.
If the user instead ends that chain with \code{collapse()}, then calling \code{compute}
generates a call to \code{compute.data_query()}. Either way the result will be
identical.
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\code{\link[=atlas_occurrences]{atlas_occurrences()}}
}
