% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute.data_request}
\alias{compute.data_request}
\alias{compute.metadata_request}
\alias{compute.files_request}
\alias{compute.query}
\title{Compute a query}
\usage{
\method{compute}{data_request}(x, ...)

\method{compute}{metadata_request}(x, ...)

\method{compute}{files_request}(x, ...)

\method{compute}{query}(x, ...)
}
\arguments{
\item{x}{An object of class \code{data_request}, \code{metadata_request} or
\code{files_request} (i.e. constructed using a pipe) or \code{query}
(i.e. constructed by \code{collapse()})}

\item{...}{Arguments passed on to other methods}
}
\value{
An object of class \code{computed_query}, which is identical to class
\code{query} except for occurrence data, where it also contains information on the
status of the request.
}
\description{
\code{compute()} is useful for several purposes. It's original
purpose is to send a request for data, which can then be processed by the
server and retrieved at a later time (via \code{collect()}).
}
