% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count.data_request}
\alias{count.data_request}
\title{Count for object of class \code{data_request}}
\usage{
\method{count}{data_request}(x, ..., wt, sort, name, type = c("record", "species"))
}
\arguments{
\item{x}{An object of class \code{data_request}, created using \code{\link[=galah_call]{galah_call()}}}

\item{...}{currently ignored}

\item{wt}{currently ignored}

\item{sort}{currently ignored}

\item{name}{currently ignored}

\item{type}{\code{string}: one of \code{c("record", "species")}. Defaults to
"record". If "species", the number of species matching the criteria will be
returned, if "record", the number of records matching the criteria will be
returned.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
\code{\link[=atlas_counts]{atlas_counts()}}, with which this function is synonymous.
}
