% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\alias{describe.data_request}
\title{Describe what fields are available}
\usage{
describe(x, ...)

\method{describe}{data_request}(x, ...)
}
\arguments{
\item{x}{An object of class \code{data_request}}

\item{...}{Other arguments, currently ignored}
}
\value{
A \code{tibble} showing the \code{id}, \code{description} and \code{data_type}
for all fields, or if \code{select()} is called, then those fields
requested by the user.
}
\description{
Unlike a local \code{tibble}, where printing to the console shows what fields
are available, when working remotely there is no way to know what fields
are present that can be queried. This function name is borrowed from SQL,
where it is used to give a read-out of fields that are in the source
database. \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
# By default, this shows all fields in the source system
galah_call() |>
  describe() |>
  collect()

# If `select()` is called, only requested fields are shown
galah_call() |>
  select(group = "basic") |>
  describe() |>
  collect()
}
}
\seealso{
\code{\link[=glimpse.data_request]{glimpse()}} for a different
way to view results of a query; \code{\link[=show_all_fields]{show_all_fields()}} for full metadata
on available fields; \code{\link[=distinct.data_request]{distinct()}}
for showing the values \emph{within} a given field.
}
