% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-distinct.R
\name{distinct.data_request}
\alias{distinct.data_request}
\title{Keep distinct/unique rows}
\usage{
\method{distinct}{data_request}(.data, ..., .keep_all = FALSE)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a
lazy data frame (e.g. from dbplyr or dtplyr). See Methods, below,
for more details.}

\item{...}{Optional variables to use when determining uniqueness. If there
are multiple rows for a given combination of inputs, only the first row
will be preserved. If omitted, will use all variables in the data frame.}

\item{.keep_all}{If \code{TRUE}, keep all variables in .data. If a combination
of \code{...} is not distinct, this keeps the first row of values.}
}
\description{
Keep only unique/distinct rows from a data frame. This is similar to
\code{\link[=unique.data.frame]{unique.data.frame()}} but considerably faster. It is evaluated lazily.
}
