% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.data_request}
\alias{filter.data_request}
\title{Filter for object of class \code{data_request}}
\usage{
\method{filter}{data_request}(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{data_request}, created using \code{\link[=galah_call]{galah_call()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
