% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_object_classes.R
\name{as_data_filter}
\alias{as_data_filter}
\alias{as_predicates_filter}
\alias{as_metadata_filter}
\alias{as_files_filter}
\alias{print.data_filter}
\alias{print.predicates_filter}
\alias{print.metadata_filter}
\alias{print.files_filter}
\title{Object classes for \code{filter()} queries}
\usage{
as_data_filter(x)

as_predicates_filter(x)

as_metadata_filter(x)

as_files_filter(x)

\method{print}{data_filter}(x, ...)

\method{print}{predicates_filter}(x, ...)

\method{print}{metadata_filter}(x, ...)

\method{print}{files_filter}(x, ...)
}
\arguments{
\item{x}{a list}
}
\description{
In galah, there are several ways to provide filter information. To ensure
these are handled and printed correctly, they are assigned classes
}
