% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_field_values.R
\name{find_field_values}
\alias{find_field_values}
\title{List valid options for a categorical field}
\usage{
find_field_values(field, limit = 20)
}
\arguments{
\item{field}{string: field to return the categories for. Use
\code{\link{find_fields}} to view valid fields.}

\item{limit}{numeric: maximum number of categories to return. 20 by default.}
}
\value{
a data.frame of field name and category name.
}
\description{
Used for checking filter values for \code{\link{ala_occurrences}} and
\code{ala_counts} are valid
}
\examples{
find_field_values("basis_of_record")
find_field_values("state")
}
