% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fields.R
\name{find_fields}
\alias{find_fields}
\title{Retrieves field names for use with data retrieval functions}
\usage{
find_fields(class = "all")
}
\arguments{
\item{class}{\code{string}: class of fields to return e.g. "Assertion". By
default all fields are returned.}
}
\value{
\code{data.frame} of fields with name, data_type, information and Darwin
Core class.
}
\description{
When building a set of filters with \code{\link{select_filters}()} or 
specifying required columns with \code{\link{select_columns}()}, this function
can be used to check that the fields map to fields stored in the ALA.
This function returns a \code{data.frame} of all fields. Field names are in
Darwin Core format, except in the case where the field is specific to the
ALA database, in which case the ALA field name is returned.
}
\examples{
\dontrun{
# Find all fields
 l <- find_fields()
 # Find assertion fields
 ass <- find_fields("assertion")
}
}
\references{
\itemize{
\item Darwin Core terms \url{https://dwc.tdwg.org/terms/}
\item ALA fields \url{https://api.ala.org.au/#ws72}
\item ALA assertion fields \url{https://api.ala.org.au/#ws81}
}
}
\seealso{
This function is used to pass valid arguments to
\code{\link{select_filters}()} & \code{\link{select_columns}()}. The information
that this function gives on spatial layers is not particularly detailed;
consider also checking \code{\link{find_layers}()} for more information.
}
