% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_fields.R
\name{find_fields}
\alias{find_fields}
\title{Retrieves a list of all field names that can be used with data
retrieval functions}
\usage{
find_fields(class = "all")
}
\arguments{
\item{class}{string: class of fields to return e.g. "Assertion"}
}
\value{
data.frame of fields with name, data_type, information and Darwin
Core class.
}
\description{
The field names are in Darwin Core format, apart from the fields
which have no DwC equivalent, such as record assertions.
}
\examples{
\dontrun{
# Find all fields
 l <- find_fields()
 # Find assertion fields
 ass <- find_fields("assertion")
}
}
\references{
\itemize{
\item Darwin Core terms \url{https://dwc.tdwg.org/terms/}
\item ALA fields \url{https://api.ala.org.au/#ws72}
\item ALA assertion fields \url{https://api.ala.org.au/#ws81}
}
}
\seealso{
\code{\link{find_layers}} to search for spatial layers
}
