% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_layers.R
\name{find_layers}
\alias{find_layers}
\title{Search ALA spatial layers}
\usage{
find_layers()
}
\value{
A \code{data.frame} of all spatial layers held by the ALA
}
\description{
The ALA stores information on a wide range of contextual and environmental
layers. \code{find_layers()} returns a \code{data.frame} of all layers
stored by the ALA.
To return layer values with occurrence records, pass the required layer
names to \code{\link{select_columns}}. To filter occurrence records by a
layer value (for layers with categorical values), pass the layer name and
value to \code{\link{select_filters}}.
}
\details{
The resulting \code{data.frame} contains the following columns:
\itemize{
 \item{layer_id: A prefix ('el' for environmental layers and 'cl' for
 contextual layers) followed by an id number. This is the value that should
 be used when referring to a layer in another function.}
 \item{name: Descriptive name of the layer.}
 \item{source_link: Link to original data source (if available)}
 \item{description: Additional details about layer}
}
}
\examples{
\dontrun{
# Find all precipitation-related layers
layers <- find_layers()
layers[grepl("Precipitation", layers$name),]

# Find the IBRA 7 layer
ibra_7_layer <- layers[grepl("IBRA 7 Regions", layers$name),]
# View values for IBRA 7 layer
find_field_values(ibra_7_layer$layer_id)
}
}
\seealso{
This function is used to pass valid arguments to
\code{\link{select_columns}()} and \code{\link{select_filters}()}.
Layers are the subset of fields that are spatially appended to each record
by the ALA; the full list is given by \code{\link{find_fields}()}.
To view valid values for a layer with categorical values, use
\code{\link{find_field_values}()}.
}
