% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ala_config.R
\name{find_reasons}
\alias{find_reasons}
\title{List valid download reasons}
\usage{
find_reasons()
}
\value{
A \code{data.frame} of valid download reasons, containing the id
and name for each reason.
}
\description{
When downloading occurrence data with \code{\link{ala_occurrences}} the 
ALA APIs require a reason for download to be specified. By default, a 
download reason of 'scientific research' is set for you, but if you wish to
change this you can do so with \code{\link{ala_config}()}. Use this function
to view the list of download reason code and names. When specifying a reason, 
you can use either the download code or name.
}
\seealso{
This function is helpful in setting up \code{\link{ala_config}()}.
}
