% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah_call.R
\name{galah_call}
\alias{galah_call}
\alias{print.data_request}
\title{Build a data request object}
\usage{
galah_call(
  taxa = NULL,
  filter = NULL,
  select = NULL,
  geolocate = NULL,
  group_by = NULL,
  ...
)

\method{print}{data_request}(object)
}
\arguments{
\item{taxa}{\code{data.frame}: generated by a call to \code{\link[=search_taxa]{search_taxa()}}. This
argument also accepts a vector of unique species identifiers.}

\item{filter}{\code{data.frame}: generated by a call to \code{\link[=select_filters]{select_filters()}}}

\item{select}{\code{data.frame}: generated by a call to \code{\link[=select_columns]{select_columns()}}}

\item{geolocate}{\code{string}: generated by a call to \code{\link[=select_locations]{select_locations()}}}

\item{...}{other function-specific request parameters}
}
\description{
Build a data request object
}
