% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.data_request}
\alias{filter.data_request}
\title{Filter for object of class \code{data_request}}
\usage{
\method{filter}{data_request}(data, ...)
}
\description{
Filter for object of class \code{data_request}
}
