% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah_group_by.R
\name{galah_group_by}
\alias{galah_group_by}
\alias{group_by.data_request}
\title{Specify fields to group when downloading record counts}
\usage{
galah_group_by(...)

\method{group_by}{data_request}(.data, ...)
}
\arguments{
\item{...}{zero or more individual column names to include}

\item{.data}{An object of class \code{data_request}}
}
\value{
If any arguments are provided, returns a \code{data.frame} with
columns \code{name} and \code{type}, as per \code{\link[=select.data_request]{select.data_request()}}.
}
\description{
\code{count.data_request()} and \code{atlas_counts()} support server-side grouping of
data. Grouping can be used to return record counts grouped by multiple, valid
fields (found by \code{search_all(fields)}).
}
\examples{
\dontrun{
galah_call() |> 
  galah_group_by(basisOfRecord) |>
  atlas_counts()
}
}
