% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galah_media.R
\name{galah_media}
\alias{galah_media}
\alias{media}
\title{Function to specify how media should be handled by \code{galah}}
\usage{
galah_media(
  ...,
  id = NULL,
  format = c("image", "sound", "video"),
  path = ".",
  filesize = c("full", "thumbnail"),
  .join = c("full_join", "inner_join", "left_join", "right_join")
)

media(
  ...,
  id = NULL,
  format = c("image", "sound", "video"),
  path = ".",
  filesize = c("full", "thumbnail"),
  .join = c("full_join", "inner_join", "left_join", "right_join")
)
}
\arguments{
\item{...}{Optional additional arguments, typically including an object of
class \code{data_request}.}

\item{id}{Optional vector giving media identifiers to be collected.}

\item{path}{String: where should data be saved?}

\item{filesize}{String: either "full" (default) or "thumbnail". Note only
applies if \code{media = "image"}.}

\item{.join}{String: if \code{type = "occurrences"}, how should occurrences and
media metadata be joined? Should be one of the \code{dplyr} mutating joins,
defaulting to \code{"full_join"}.}

\item{media}{Vector: One or more of "image", "sound" or "video", defaulting
to all three. Ignored if \code{id} is specified.}
}
\description{
In prep
}
