% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by.data_request}
\alias{group_by.data_request}
\title{Group by for object of class \code{data_request}}
\usage{
group_by.data_request(.data, ...)
}
\arguments{
\item{.data}{An object of class \code{data_request}, created using \code{\link[=galah_call]{galah_call()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
