% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koala.R
\docType{package}
\name{koala}
\alias{koala}
\title{\pkg{koala}}
\description{
This project enables the R community to access data and tools hosted by the
Atlas of Living Australia. The goal of the project is
to enable basic species and related information to be queried and output in
standard formats for R. koala is based around the extensive web services
provided by the Atlas; see the API link below.
}
\details{
If you have any questions or suggestions, please email support@ala.org.au.
}
\section{Functions}{

\strong{Data functions}
\itemize{
  \item\code{\link{ala_taxa}} Taxon information
  \item\code{\link{ala_occurrences}} Download occurrence records
  \item\code{\link{ala_counts}} Summary statistics for occurrence records
  \item\code{\link{ala_media}} Download images and sounds
}
\strong{Helper functions}
\itemize{
  \item\code{\link{ala_config}} Package configuration options
  \item\code{\link{ala_citation}} Citation for a dataset
}
\strong{Select_ functions}
\itemize{
  \item\code{\link{select_filters}} Filter records
  \item\code{\link{select_locations}} Specify location
  \item\code{\link{select_columns}} Columns to return in occurrence download
}
\strong{Find_ functions}
\itemize{
  \item\code{\link{find_profiles}} List data quality profiles
  \item\code{\link{find_profile_attributes}} List filters included in a data quality profile
  \item\code{\link{find_fields}} List occurrence record fields
  \item\code{\link{find_field_values}} List possible values for a given field
  \item\code{\link{find_layers}} List spatial layers
  \item\code{\link{find_reasons}} List valid download reasons
}
}

\references{
\url{https://api.ala.org.au/}
}
