% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{identify}
\alias{filter}
\alias{select}
\alias{group_by}
\alias{slice_head}
\alias{count}
\alias{st_crop}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{count}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:slice]{slice_head}}}

  \item{graphics}{\code{\link[graphics]{identify}}}

  \item{sf}{\code{\link[sf]{st_crop}}}
}}

