% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_minifunctions.R
\name{search_minifunctions}
\alias{search_minifunctions}
\alias{search_atlases}
\alias{search_reasons}
\alias{search_ranks}
\alias{search_fields}
\alias{search_values}
\alias{search_assertions}
\alias{search_profiles}
\alias{search_profile_attributes}
\alias{search_providers}
\alias{search_collections}
\alias{search_datasets}
\alias{search_licences}
\title{Search for a valid option within various types of information}
\usage{
search_assertions(query)

search_atlases(query)

search_collections(query)

search_datasets(query)

search_providers(query)

search_fields(query)

search_licences(query)

search_reasons(query)

search_ranks(query)

search_profiles(query)
}
\arguments{
\item{query}{: A string specifying a search term. Searches are not
case-sensitive.}
}
\value{
An object of class \code{tbl_df} and \code{data.frame} (aka a tibble)
}
\description{
The living atlases store a huge amount of information, above and beyond the
occurrence records that are their main output. In \code{galah}, one way that
users can investigate this information is by searching for a specific option
or category for the type of information they are interested in.
Functions prefixed with \code{search_} do this, displaying any matches to a
search term within the valid options for the information specified by the
suffix.
}
\details{
There are five categories of information, each with their own
sub-functions, or "minifunctions" for to display each type.
The available \code{search_} sub-functions are:\tabular{lll}{
   \strong{Category} \tab \strong{Sub-function} \tab \strong{Description} \cr
   configuration \tab \code{search_atlases()} \tab Search for what atlases are available \cr
    \tab \code{search_apis()} \tab Search for what APIs & functions are available for each atlas \cr
    \tab \code{search_reasons()} \tab Show what values are acceptable as 'download reasons' for a specified atlas \cr
   taxonomy \tab \code{search_taxa()} \tab Search for one or more taxonomic names \cr
    \tab \code{search_identifiers()} \tab Take a universal identifier and return taxonomic information \cr
    \tab \code{search_ranks()} \tab Search for valid taxonomic ranks (e.g. Kingdom, Class, Order, etc.) \cr
   filters \tab \code{search_fields()} \tab Search for fields that are stored in an atlas \cr
    \tab \code{search_assertions()} \tab Search for results of data quality checks run by each atlas \cr
    \tab \code{search_licences()} \tab Search for copyright licences applied to media \cr
   group filters \tab \code{search_profiles()} \tab Search for what data profiles are available \cr
    \tab \code{search_lists()} \tab Search for what species lists are available \cr
   data providers \tab \code{search_providers()} \tab Search for which institutions have provided data \cr
    \tab \code{search_collections()} \tab Search for the specific collections within those institutions \cr
    \tab \code{search_datasets()} \tab Search for the data groupings within those collections \cr
}
}
\section{Examples}{


Search using a single term

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_taxa("Reptilia")
#> # A tibble: 1 x 9
#>   search_term scientific_name taxon_~1 rank  match~2 kingdom phylum class issues
#>   <chr>       <chr>           <chr>    <chr> <chr>   <chr>   <chr>  <chr> <chr> 
#> 1 Reptilia    REPTILIA        https:/~ class exactM~ Animal~ Chord~ Rept~ noIss~
#> # ... with abbreviated variable names 1: taxon_concept_id, 2: match_type
}\if{html}{\out{</div>}}

Note that \code{search_taxa()} is not case sensitive

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_taxa("reptilia") # not case sensitive
#> # A tibble: 1 x 9
#>   search_term scientific_name taxon_~1 rank  match~2 kingdom phylum class issues
#>   <chr>       <chr>           <chr>    <chr> <chr>   <chr>   <chr>  <chr> <chr> 
#> 1 reptilia    REPTILIA        https:/~ class exactM~ Animal~ Chord~ Rept~ noIss~
#> # ... with abbreviated variable names 1: taxon_concept_id, 2: match_type
}\if{html}{\out{</div>}}

Search multiple taxa. \code{search_taxa()} will return one row per taxon

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_taxa(c("reptilia", "mammalia"))
#> # A tibble: 2 x 10
#>   search_term scient~1 taxon~2 rank  match~3 kingdom phylum class issues verna~4
#>   <chr>       <chr>    <chr>   <chr> <chr>   <chr>   <chr>  <chr> <chr>  <chr>  
#> 1 reptilia    REPTILIA https:~ class exactM~ Animal~ Chord~ Rept~ noIss~ <NA>   
#> 2 mammalia    MAMMALIA https:~ class exactM~ Animal~ Chord~ Mamm~ noIss~ Mammals
#> # ... with abbreviated variable names 1: scientific_name, 2: taxon_concept_id,
#> #   3: match_type, 4: vernacular_name
}\if{html}{\out{</div>}}

\code{galah_identify()} uses \code{search_taxa()} to narrow data queries

Look up a unique taxon identifier

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_identifiers(identifier = "https://id.biodiversity.org.au/node/apni/2914510")
#> # A tibble: 1 x 14
#>   scient~1 scien~2 taxon~3 rank  match~4 kingdom phylum class order family genus
#>   <chr>    <chr>   <chr>   <chr> <chr>   <chr>   <chr>  <chr> <chr> <chr>  <chr>
#> 1 Eucalyp~ Maiden  https:~ spec~ taxonI~ Plantae Charo~ Equi~ Myrt~ Myrta~ Euca~
#> # ... with 3 more variables: species <chr>, vernacular_name <chr>,
#> #   issues <chr>, and abbreviated variable names 1: scientific_name,
#> #   2: scientific_name_authorship, 3: taxon_concept_id, 4: match_type
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{galah_call() |>
  galah_identify("reptilia") |>
  atlas_counts()
#> # A tibble: 1 x 1
#>     count
#>     <int>
#> 1 1462278
}\if{html}{\out{</div>}}

Search for all fields that use include the word "date"

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("date")
#> # A tibble: 22 x 4
#>   id             description                                         type  link 
#>   <chr>          <chr>                                               <chr> <chr>
#> 1 dateIdentified http://rs.tdwg.org/dwc/terms/dateIdentified         fiel~ <NA> 
#> 2 datePrecision  The precision of the date information for the reco~ fiel~ <NA> 
#> 3 eventDate      The ISO formatted date of observation, specimen co~ fiel~ <NA> 
#> 4 eventDateEnd   <NA>                                                fiel~ <NA> 
#> # ... with 18 more rows
}\if{html}{\out{</div>}}

Search for all fields with the string "basisofrecord"

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("basisofrecord")
#> # A tibble: 2 x 4
#>   id                description                                      type  link 
#>   <chr>             <chr>                                            <chr> <chr>
#> 1 basisOfRecord     What this is a record of e.g. specimen, human o~ fiel~ <NA> 
#> 2 raw_basisOfRecord The basis of record as supplied by the data pub~ fiel~ <NA>
}\if{html}{\out{</div>}}

Search for all fields that have information for "marine"

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("marine") |> 
  head() # only show first 5 results
#> # A tibble: 6 x 4
#>   id      description                                                type  link 
#>   <chr>   <chr>                                                      <chr> <chr>
#> 1 biome   "A field indicating if the record has been located in Ter~ fiel~  <NA>
#> 2 cl2105  "FAO Fishery Statistical Areas FAO Fishing areas (Marine)" laye~ "htt~
#> 3 cl10948 "CAPAD 2018 Terrestrial The Collaborative Australian Prot~ laye~ "htt~
#> 4 cl10957 "CAPAD 2018 Marine The Collaborative Australian Protected~ laye~ "htt~
#> 5 el1056  "Endemism (Non-marine) Endemism (Non-marine) based on Cri~ laye~ ""   
#> 6 el957   "Averaged Topographic Relief  This data represents the av~ laye~ "htt~
}\if{html}{\out{</div>}}

Search for all Wordclim layers

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("worldclim")
#> # A tibble: 38 x 4
#>   id      description                                                type  link 
#>   <chr>   <chr>                                                      <chr> <chr>
#> 1 el10982 WorldClim 2.1: Temperature - warmest month max Max Temper~ laye~ http~
#> 2 el10981 WorldClim 2.1: Temperature - seasonality Temperature Seas~ laye~ http~
#> 3 el10980 WorldClim 2.1: Temperature - isothermality Isothermality   laye~ http~
#> 4 el10990 WorldClim 2.1: Precipitation - wettest month Precipitatio~ laye~ http~
#> # ... with 34 more rows
}\if{html}{\out{</div>}}
}

\seealso{
\code{\link[=show_all]{show_all()}} and \code{\link[=search_all]{search_all()}} are flexible helper functions to
show and/or search for information returned by each look-up sub-function.
}
