% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fields.R, R/search_identifiers.R,
%   R/search_minifunctions.R, R/show_all_assertions.R, R/show_all_atlases.R,
%   R/show_all_collectory.R, R/show_all_licences.R, R/show_all_profiles.R,
%   R/show_all_ranks.R, R/show_all_reasons.R
\name{search_fields}
\alias{search_fields}
\alias{search_identifiers}
\alias{search_minifunctions}
\alias{search_atlases}
\alias{search_reasons}
\alias{search_taxa}
\alias{search_ranks}
\alias{search_values}
\alias{search_assertions}
\alias{search_profiles}
\alias{search_profile_attributes}
\alias{search_providers}
\alias{search_collections}
\alias{search_datasets}
\alias{search_licences}
\title{Functions for searching detailed metadata}
\usage{
search_fields(query)

search_identifiers(identifier)

search_assertions(query)

search_atlases(query)

search_collections(query)

search_datasets(query)

search_providers(query)

search_licences(query)

search_profiles(query)

search_ranks(query)

search_reasons(query)
}
\arguments{
\item{identifier}{\code{string}: A vector containing one or more taxonomic
identifiers, given as strings.}
}
\description{
Some descriptive text
}
\section{Examples}{


Search using a single term

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_taxa("Reptilia")
#> # A tibble: 1 x 9
#>   search_term scientific_name taxon_~1 rank  match~2 kingdom phylum class issues
#>   <chr>       <chr>           <chr>    <chr> <chr>   <chr>   <chr>  <chr> <chr> 
#> 1 Reptilia    REPTILIA        https:/~ class exactM~ Animal~ Chord~ Rept~ noIss~
#> # ... with abbreviated variable names 1: taxon_concept_id, 2: match_type
}\if{html}{\out{</div>}}

Note that \code{search_taxa()} is not case sensitive

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_taxa("reptilia") # not case sensitive
#> # A tibble: 1 x 9
#>   search_term scientific_name taxon_~1 rank  match~2 kingdom phylum class issues
#>   <chr>       <chr>           <chr>    <chr> <chr>   <chr>   <chr>  <chr> <chr> 
#> 1 reptilia    REPTILIA        https:/~ class exactM~ Animal~ Chord~ Rept~ noIss~
#> # ... with abbreviated variable names 1: taxon_concept_id, 2: match_type
}\if{html}{\out{</div>}}

Search multiple taxa. \code{search_taxa()} will return one row per taxon

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_taxa(c("reptilia", "mammalia"))
#> # A tibble: 2 x 10
#>   search_term scient~1 taxon~2 rank  match~3 kingdom phylum class issues verna~4
#>   <chr>       <chr>    <chr>   <chr> <chr>   <chr>   <chr>  <chr> <chr>  <chr>  
#> 1 reptilia    REPTILIA https:~ class exactM~ Animal~ Chord~ Rept~ noIss~ <NA>   
#> 2 mammalia    MAMMALIA https:~ class exactM~ Animal~ Chord~ Mamm~ noIss~ Mammals
#> # ... with abbreviated variable names 1: scientific_name, 2: taxon_concept_id,
#> #   3: match_type, 4: vernacular_name
}\if{html}{\out{</div>}}

\code{galah_identify()} uses \code{search_taxa()} to narrow data queries

Look up a unique taxon identifier

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_identifiers(identifier = "https://id.biodiversity.org.au/node/apni/2914510")
#> # A tibble: 1 x 14
#>   scient~1 scien~2 taxon~3 rank  match~4 kingdom phylum class order family genus
#>   <chr>    <chr>   <chr>   <chr> <chr>   <chr>   <chr>  <chr> <chr> <chr>  <chr>
#> 1 Eucalyp~ Maiden  https:~ spec~ taxonI~ Plantae Charo~ Equi~ Myrt~ Myrta~ Euca~
#> # ... with 3 more variables: species <chr>, vernacular_name <chr>,
#> #   issues <chr>, and abbreviated variable names 1: scientific_name,
#> #   2: scientific_name_authorship, 3: taxon_concept_id, 4: match_type
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode r">}}\preformatted{galah_call() |>
  galah_identify("reptilia") |>
  atlas_counts()
#> # A tibble: 1 x 1
#>     count
#>     <int>
#> 1 1458849
}\if{html}{\out{</div>}}

Search for all fields that use include the word "date"

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("date")
#> # A tibble: 19 x 3
#>    id                          description                                 type 
#>    <chr>                       <chr>                                       <chr>
#>  1 dateIdentified              http://rs.tdwg.org/dwc/terms/dateIdentified fiel~
#>  2 datePrecision               The precision of the date information for ~ fiel~
#>  3 eventDate                   The ISO formatted date of observation, spe~ fiel~
#>  4 eventDateEnd                <NA>                                        fiel~
#>  5 firstLoadedDate             <NA>                                        fiel~
#>  6 georeferencedDate           http://rs.tdwg.org/dwc/terms/georeferenced~ fiel~
#>  7 lastAssertionDate           <NA>                                        fiel~
#>  8 lastLoadDate                <NA>                                        fiel~
#>  9 lastProcessedDate           <NA>                                        fiel~
#> 10 measurementDeterminedDate   http://rs.tdwg.org/dwc/terms/measurementDe~ fiel~
#> 11 month                       Month of observation, specimen collection ~ fiel~
#> 12 raw_dateIdentified          <NA>                                        fiel~
#> 13 raw_datePrecision           <NA>                                        fiel~
#> 14 raw_eventDate               <NA>                                        fiel~
#> 15 raw_georeferencedDate       <NA>                                        fiel~
#> 16 raw_lastLoadDate            <NA>                                        fiel~
#> 17 raw_lastProcessedDate       <NA>                                        fiel~
#> 18 relationshipEstablishedDate http://rs.tdwg.org/dwc/terms/relationshipE~ fiel~
#> 19 verbatimEventDate           http://rs.tdwg.org/dwc/terms/verbatimEvent~ fiel~
}\if{html}{\out{</div>}}

Search for all fields with the string "basisofrecord"

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("basisofrecord")
#> # A tibble: 2 x 3
#>   id                description                                            type 
#>   <chr>             <chr>                                                  <chr>
#> 1 basisOfRecord     What this is a record of e.g. specimen, human observa~ fiel~
#> 2 raw_basisOfRecord The basis of record as supplied by the data publisher~ fiel~
}\if{html}{\out{</div>}}

Search for all fields that have information for "marine"

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("marine") |> 
  head() # only show first 5 results
#> # A tibble: 1 x 3
#>   id    description                                                        type 
#>   <chr> <chr>                                                              <chr>
#> 1 biome "A field indicating if the record has been located in Terrestrial~ fiel~
}\if{html}{\out{</div>}}

Search for all Wordclim layers

\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("worldclim")
#> # A tibble: 0 x 3
#> # ... with 3 variables: id <chr>, description <chr>, type <chr>
}\if{html}{\out{</div>}}
}

