% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_fields.R, R/search_identifiers.R,
%   R/search_minifunctions.R, R/search_values.R, R/show_all_assertions.R,
%   R/show_all_atlases.R, R/show_all_collectory.R, R/show_all_licences.R,
%   R/show_all_profiles.R, R/show_all_ranks.R, R/show_all_reasons.R
\name{search_fields}
\alias{search_fields}
\alias{search_identifiers}
\alias{search_minifunctions}
\alias{search_atlases}
\alias{search_reasons}
\alias{search_taxa}
\alias{search_ranks}
\alias{search_values}
\alias{search_assertions}
\alias{search_profiles}
\alias{search_profile_attributes}
\alias{search_providers}
\alias{search_collections}
\alias{search_datasets}
\alias{search_licences}
\title{Functions for searching detailed metadata}
\usage{
search_fields(query)

search_identifiers(identifier)

search_values(field, query)

search_assertions(query)

search_atlases(query)

search_collections(query)

search_datasets(query)

search_providers(query)

search_licences(query)

search_profiles(query)

search_ranks(query)

search_reasons(query)
}
\arguments{
\item{query}{\code{string}: A search string. Not case sensitive.}

\item{identifier}{\code{string}: A vector containing one or more taxonomic
identifiers, given as strings.}

\item{field}{\code{string}: field to return the categories for. Use
\code{\link[=search_fields]{search_fields()}} to view valid fields.}
}
\description{
Some descriptive text
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode {r, child = "man/rmd/setup.Rmd"}">}}\preformatted{}\if{html}{\out{</div>}}

Search using a single term

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{search_taxa("Reptilia")
}\if{html}{\out{</div>}}

Note that \code{search_taxa()} is not case sensitive

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{search_taxa("reptilia") # not case sensitive
}\if{html}{\out{</div>}}

Search multiple taxa. \code{search_taxa()} will return one row per taxon

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{search_taxa(c("reptilia", "mammalia"))
}\if{html}{\out{</div>}}

\code{galah_identify()} uses \code{search_taxa()} to narrow data queries

Look up a unique taxon identifier


\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_identifiers(identifier = "https://id.biodiversity.org.au/node/apni/2914510")
#> # A tibble: 1 x 14
#>   scient~1 scien~2 taxon~3 rank  match~4 kingdom phylum class order family genus
#>   <chr>    <chr>   <chr>   <chr> <chr>   <chr>   <chr>  <chr> <chr> <chr>  <chr>
#> 1 Eucalyp~ Maiden  https:~ spec~ taxonI~ Plantae Charo~ Equi~ Myrt~ Myrta~ Euca~
#> # ... with 3 more variables: species <chr>, vernacular_name <chr>,
#> #   issues <chr>, and abbreviated variable names 1: scientific_name,
#> #   2: scientific_name_authorship, 3: taxon_concept_id, 4: match_type
}\if{html}{\out{</div>}}

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{galah_call() |>
  galah_identify("reptilia") |>
  atlas_counts()
#> # A tibble: 1 x 1
#>     count
#>     <int>
#> 1 1458841
}\if{html}{\out{</div>}}

Search for all fields that use include the word "date"
\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("date")
#> # A tibble: 22 x 4
#>   id             description                                         type  link 
#>   <chr>          <chr>                                               <chr> <chr>
#> 1 dateIdentified http://rs.tdwg.org/dwc/terms/dateIdentified         fiel~ <NA> 
#> 2 datePrecision  The precision of the date information for the reco~ fiel~ <NA> 
#> 3 eventDate      The ISO formatted date of observation, specimen co~ fiel~ <NA> 
#> 4 eventDateEnd   <NA>                                                fiel~ <NA> 
#> # ... with 18 more rows
}\if{html}{\out{</div>}}

Search for all fields with the string "basisofrecord"

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{search_fields("basisofrecord")
}\if{html}{\out{</div>}}

Search for all fields that have information for "marine"

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{search_fields("marine") |> 
  head() # only show first 5 results
}\if{html}{\out{</div>}}

Search for all Wordclim layers
\if{html}{\out{<div class="sourceCode r">}}\preformatted{search_fields("worldclim")
#> # A tibble: 38 x 4
#>   id      description                                                type  link 
#>   <chr>   <chr>                                                      <chr> <chr>
#> 1 el10982 WorldClim 2.1: Temperature - warmest month max Max Temper~ laye~ http~
#> 2 el10981 WorldClim 2.1: Temperature - seasonality Temperature Seas~ laye~ http~
#> 3 el10980 WorldClim 2.1: Temperature - isothermality Isothermality   laye~ http~
#> 4 el10990 WorldClim 2.1: Precipitation - wettest month Precipitatio~ laye~ http~
#> # ... with 34 more rows
}\if{html}{\out{</div>}}

To find all the data quality arguments used in the profile "CSDM"

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{search_profile_attributes("CSDM")
}\if{html}{\out{</div>}}

Then get a free-text description of each filter used in the "CSDM" profile

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{profile_info <- search_profile_attributes("CSDM")
profile_info$description
}\if{html}{\out{</div>}}
}

