% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_taxa.R
\name{search_taxa}
\alias{search_taxa}
\title{Search taxon information}
\usage{
search_taxa(query, downto = NULL, include_ids = FALSE)
}
\arguments{
\item{query}{\code{string}: A vector containing one or more search terms,
given as strings. Search terms can be scientific or common names, or
taxanomic identifiers. If greater control is required to disambiguate search
terms, taxonomic levels can be provided explicitly via a named \code{list}
for a single name. See vignette(taxonomic_information for using search_taxa
for multiple taxons)
Note that searches are not case-sensitive.}

\item{downto}{\code{string}: A taxonomic rank to search down to}

\item{include_ids}{\code{logical} Include unique taxonomic identifiers for
all of the ranks? \code{FALSE} by default.}
}
\description{
In the ALA, all records are associated with an identifier that uniquely
identifies the taxon to which that record belongs. However, taxonomic names
can be ambiguous due to homonymy; i.e. re-use of names (common or
scientific) in different clades. Hence, \code{search_taxa} provides a means
to search for taxonomic names and check the results are 'correct' before
proceeding to download data via \code{\link{ala_occurrences}()},
\code{\link{ala_species}()} or \code{\link{ala_counts}()}.
}
