% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_taxa.R
\name{search_taxa}
\alias{search_taxa}
\alias{search_taxa.data_request}
\alias{search_taxa.default}
\title{Taxon information}
\usage{
search_taxa(...)

\method{search_taxa}{data_request}(request, ...)

\method{search_taxa}{default}(query)
}
\arguments{
\item{query}{\code{string}: A vector containing one or more search terms,
given as strings. Search terms can be scientific or common names, or
taxonomic identifiers. If greater control is required to disambiguate search
terms, taxonomic levels can be provided explicitly via a named \code{list}
for a single name, or a \code{data.frame} for multiple names (see examples).
Note that searches are not case-sensitive.}
}
\value{
An object of class \code{tbl_df}, \code{data.frame} (aka a tibble) and \code{ala_id}
containing taxonomic information.
}
\description{
In the ALA, all records are associated with an identifier that uniquely
identifies the taxon to which that record belongs. However, taxonomic names
are often ambiguous due to homonymy; i.e. re-use of names (common or
scientific) in different clades. Hence, \code{search_taxa} provides a means
to search for taxonomic names and check the results are 'correct' before
proceeded to download data via \code{\link[=atlas_occurrences]{atlas_occurrences()}},
\code{\link[=atlas_species]{atlas_species()}} or \code{\link[=atlas_counts]{atlas_counts()}}. The resulting
\code{data.frame} of taxonomic information can be passed directly to
\code{atlas_} functions to filter records to the specified taxon or taxa.
}
\examples{
# Search using a single term
search_taxa("Reptilia")
# or equivalently:
search_taxa("reptilia") # not case sensitive

# Search multiple taxa
search_taxa(c("reptilia", "mammalia")) # returns one row per taxon

}
\seealso{
\code{\link[=find_taxa]{find_taxa()}} for how to get names if taxonomic identifiers are
already known. \code{\link[=galah_select]{galah_select()}}, \code{\link[=galah_filter]{galah_filter()}} and
\code{\link[=galah_geolocate]{galah_geolocate()}} for other ways to restrict the information returned
by \code{\link[=atlas_occurrences]{atlas_occurrences()}} and related functions.
\code{\link[=atlas_taxonomy]{atlas_taxonomy()}} to look up taxonomic trees.
}
