% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.data_request}
\alias{select.data_request}
\title{Select for object of class \code{data_request}}
\usage{
\method{select}{data_request}(.data, ..., group = c("basic", "event", "media", "assertions"))
}
\arguments{
\item{.data}{An object of class \code{data_request}, created using \code{\link[=galah_call]{galah_call()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
