% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_columns.R
\name{select_columns}
\alias{select_columns}
\title{Build dataframe of columns to keep}
\usage{
select_columns(..., group)
}
\arguments{
\item{...}{individual column names to return}

\item{group}{string: name of column group to include (either \code{'basic'}
or \code{'event'})}
}
\description{
The result of `select_columns` can be passed to the `columns` argument in
\code{\link{ala_occurrences}}.
}
\details{
Calling the argument \code{group = 'basic'} returns the following columns:
\itemize{
  \item\code{decimalLatitude}
  \item\code{decimalLongitude}
  \item\code{eventDate}
  \item\code{taxon_name}
  \item\code{taxonConceptID}
  \item\code{recordID}
  \item\code{data_resource}
}
Conversely, using \code{group = 'event'} returns the following columns:
\itemize{
  \item\code{eventRemarks}
  \item\code{eventTime}
  \item\code{eventID}
  \item\code{eventDate}
  \item\code{samplingEffort}
  \item\code{samplingProtocol}
}
}
