% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_filters.R
\name{select_filters}
\alias{select_filters}
\title{Select filters to narrow down occurrence queries}
\usage{
select_filters(..., profile = NULL)
}
\arguments{
\item{...}{filters, in the form \code{field = value}}

\item{profile}{string: (optional) a data quality profile to apply to the
records. See \code{\link{find_profiles}} for valid profiles. By default
no profile is applied.}
}
\value{
data.frame of filter values
}
\description{
The result of \code{select_filters} can be passed to the `filters`
argument in \code{\link{ala_occurrences}} and \code{\link{ala_counts}}
}
\examples{
# Create custom filters with the default ALA data quality profile
filters <- select_filters(basisOfRecord = "HumanObservation", year = 2020,
            stateProvince = "New South Wales", profile = "ALA")

}
