% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_filters.R
\name{select_filters}
\alias{select_filters}
\title{Build a dataframe from a filter list}
\usage{
select_filters(filters = NULL, data_quality_profile = NULL)
}
\arguments{
\item{filters}{string: a list to narrow down the search, in the form
`list(field = value)`.}

\item{data_quality_profile}{string: a data quality profile to apply to the
records. See `ala_data_profiles()` for valid profiles. By default, no
profile is applied.}
}
\value{
dataframe of filter values
}
\description{
Build a dataframe from a filter list
}
