% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_assertions.R, R/show_all_atlases.R,
%   R/show_all_cached_files.R, R/show_all_collectory.R, R/show_all_fields.R,
%   R/show_all_licences.R, R/show_all_media.R, R/show_all_minifunctions.R,
%   R/show_all_profiles.R, R/show_all_ranks.R, R/show_all_reasons.R
\name{show_all_assertions}
\alias{show_all_assertions}
\alias{show_all_atlases}
\alias{show_all_cached_files}
\alias{show_all_collections}
\alias{show_all_datasets}
\alias{show_all_providers}
\alias{show_all_fields}
\alias{show_all_licences}
\alias{show_all_media}
\alias{show_all_minifunctions}
\alias{show_all_reasons}
\alias{show_all_ranks}
\alias{show_all_profiles}
\alias{show_all_values}
\title{Functions for showing detailed metadata}
\usage{
show_all_assertions()

show_all_atlases()

show_all_cached_files()

show_all_collections()

show_all_datasets()

show_all_providers()

show_all_fields()

show_all_licences()

show_all_media(df)

show_all_profiles()

show_all_ranks()

show_all_reasons()
}
\arguments{
\item{ids}{either a \code{tibble} from \code{atlas_occurrences} or a vector giving media ids (parsed or unparsed)}
}
\value{
An object of class \code{tbl_df} and \code{data.frame} (aka a tibble)
}
\description{
Some text
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode {r, child = "man/rmd/setup.Rmd"}">}}\preformatted{}\if{html}{\out{</div>}}

See all supported atlases

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{show_all_atlases()
}\if{html}{\out{</div>}}

See a listing of all valid fields and layers


\if{html}{\out{<div class="sourceCode r">}}\preformatted{show_all_fields()
#> # A tibble: 813 x 4
#>   id                  description                                    type  link 
#>   <chr>               <chr>                                          <chr> <chr>
#> 1 abcdTypeStatus      ABCD field in use by herbaria                  fiel~ <NA> 
#> 2 acceptedNameUsage   http://rs.tdwg.org/dwc/terms/acceptedNameUsage fiel~ <NA> 
#> 3 acceptedNameUsageID http://rs.tdwg.org/dwc/terms/acceptedNameUsag~ fiel~ <NA> 
#> 4 accessRights        <NA>                                           fiel~ <NA> 
#> # ... with 809 more rows
}\if{html}{\out{</div>}}

Show a listing of all accepted reasons for downloading occurrence data

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{show_all_reasons()
}\if{html}{\out{</div>}}

Add your download reason when configuring your session with \code{\link[=galah_config]{galah_config()}}

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE, eval = FALSE}">}}\preformatted{galah_config(download_reason_id = 3)
}\if{html}{\out{</div>}}

Show a list of all available data quality profiles

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{show_all_profiles()
}\if{html}{\out{</div>}}

Values in the \code{shortName} column can be used with \code{\link[=galah_filter]{galah_filter()}} to
narrow your data query results

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{galah_filter(profile == "ALA")
}\if{html}{\out{</div>}}

Show a listing of all taxonomic ranks

\if{html}{\out{<div class="sourceCode r">}}\preformatted{show_all_ranks()
#> # A tibble: 69 x 2
#>      id name        
#>   <int> <chr>       
#> 1     1 root        
#> 2     2 superkingdom
#> 3     3 kingdom     
#> 4     4 subkingdom  
#> # ... with 65 more rows
}\if{html}{\out{</div>}}

Use ranks with \code{\link[=galah_down_to]{galah_down_to()}} and \code{\link[=atlas_taxonomy]{atlas_taxonomy()}} to get taxonomic
trees

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE}">}}\preformatted{galah_call() \%>\%
  galah_identify("fungi") \%>\%
  galah_down_to(subphylum) \%>\%
  atlas_taxonomy()
}\if{html}{\out{</div>}}

Configure caching and create a query to cache with \code{\link[=galah_config]{galah_config()}}

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE, eval = FALSE}">}}\preformatted{galah_config(caching = TRUE)
dat <- atlas_counts(group_by = galah_group_by(year))
}\if{html}{\out{</div>}}

Show a listing of previously cached files

\if{html}{\out{<div class="sourceCode {r, comment = "#>", collapse = TRUE, eval = FALSE}">}}\preformatted{show_all_cached_files()
}\if{html}{\out{</div>}}
}

