% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_all_ranks.R
\name{show_all_ranks}
\alias{show_all_ranks}
\title{Find valid taxonomic ranks}
\usage{
show_all_ranks()
}
\value{
An object of class \code{tbl_df} and \code{data.frame} (aka a tibble) of
available ranks
}
\description{
Return taxonomic ranks recognised by the ALA.
}
\seealso{
This function provides a
reference that is useful when specifying the \code{down_to} argument of
\code{\link[=atlas_taxonomy]{atlas_taxonomy()}}.
}
