% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_field_values.R, R/show_list_values.R,
%   R/show_profile_values.R, R/show_values.R
\name{show_field_values}
\alias{show_field_values}
\alias{search_field_values}
\alias{show_list_values}
\alias{show_profile_values}
\alias{show_values}
\title{Show values}
\usage{
show_field_values(field)

search_field_values(field, query)

show_list_values(list)

show_profile_values(profile)

show_values(df, type, entry)
}
\arguments{
\item{field}{\code{string}: field to return the categories for. Use
\code{\link[=search_fields]{search_fields()}} to view valid fields.}

\item{query}{\code{string}: A search string. Not case sensitive.}

\item{list}{\code{string}: list to return information from}

\item{profile}{\code{string}: a data quality profile name, short name or id.
See \code{\link[=show_all_profiles]{show_all_profiles()}} for valid filters}
}
\description{
In early development
}
