% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_values.R
\name{show_values}
\alias{show_values}
\alias{search_values}
\title{Show values of a specific field}
\usage{
show_values(df)

search_values(df, query)
}
\arguments{
\item{df}{A search result from \code{\link[=search_fields]{search_fields()}}, \code{\link[=search_profiles]{search_profiles()}} or
\code{\link[=search_lists]{search_lists()}}.}

\item{query}{A string specifying a search term. Not case sensitive.}

\item{type}{The type of information, automatically generated from the
\code{search_} query.}

\item{entry}{The query searched, automatically generated from the \code{search_} query}

\item{field}{A string specifying a field to return the categories for. Use
\code{\link[=search_fields]{search_fields()}} to view valid fields.}
}
\value{
A \code{tibble} of values
}
\description{
Users may wish to see the specific values \emph{within} a chosen field, profile
or list to narrow queries or understand more about the information of
interest. \code{show_values()} provides users with these values.
}
\details{
Each \strong{Field} contains categorical or numeric values.
For example:
\itemize{
\item The \code{field} "year" contains values 2021, 2020, 2019, etc.
\item The \code{field} "stateProvince" contains values New South Wales, Victoria, Queensland, etc.
These are used to narrow queries with \code{\link[=galah_filter]{galah_filter()}}.
}

Each \strong{Profile} contains individual filters.
For example, the "ALA" profile filter values include:
\itemize{
\item Exclude all records where spatial validity is FALSE
\item Exclude all records with a latitude value of zero
\item Exclude all records with a longitude value of zero
}

Each \strong{List} contains list items (usually taxonomic names).
For example, the Endangered Plant species list values include:
\itemize{
\item Acacia curranii (Curly-bark Wattle)
\item Brachyscome papillosa (Mossgiel Daisy)
\item Solanum karsense (Menindee Nightshade)
}
}
\examples{
search_fields("cl22") |> show_values()

# See individual filters within data quality profile "ALA"
search_profiles("ALA") |> show_values()

# See items within species list "dr19257"
search_lists("dr19257") |> show_values()

}
