% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_head.R
\name{slice_head.data_request}
\alias{slice_head.data_request}
\title{Slice for object of class \code{data_request}}
\usage{
\method{slice_head}{data_request}(.data, ..., n, prop)
}
\arguments{
\item{.data}{An object of class \code{data_request}, created using \code{\link[=galah_call]{galah_call()}}}

\item{...}{currently ignored}

\item{n}{The number of rows to be returned. If data are grouped
(using \link{group_by}), this operation will be performed on each group.}

\item{prop}{currently ignored}
}
\value{
An updated object of class \code{data_request}, whose \code{limit} argument
is set with an integer value provided by \code{n}.
}
\description{
This is a simple function to set the 'limit' argument in \code{\link[=atlas_counts]{atlas_counts()}}
using \code{dplyr} syntax.
}
