% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-summarise.R
\name{summarise.data_request}
\alias{summarise.data_request}
\alias{summarize.data_request}
\title{Summarise each group down to one row}
\usage{
summarise.data_request(.data, ...)

summarize.data_request(.data, ...)
}
\arguments{
\item{...}{Name-value pairs of summary functions. The name will be the
name of the variable in the result. The value can be a single function
such as \code{min(x)}, \code{n()}, or \code{sum()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{summarise()} creates a new data frame. It returns one row for each combination
of grouping variables; if there are no grouping variables, the output will
have a single row summarising all observations in the input. It will contain
one column for each grouping variable and one column for each of the summary
statistics that you have specified.
}
\details{
Like all \code{dplyr} extensions in \code{galah}, this function amends a \code{data_request},
\code{metadata_request} or \code{files_request}, and is evaluated lazily.

\code{summarise()} and \code{summarize()} are synonyms.
}
