% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_profile.R, R/galah_filter.R,
%   R/geolocate.R, R/galah_group_by.R, R/graphics-identify.R, R/galah_select.R,
%   R/superseded-functions.R
\name{galah_apply_profile}
\alias{galah_apply_profile}
\alias{galah_filter}
\alias{galah_geolocate}
\alias{galah_group_by}
\alias{galah_identify}
\alias{galah_select}
\alias{superseded_functions}
\title{Superseded functions}
\usage{
galah_apply_profile(...)

galah_filter(...)

galah_geolocate(..., type = c("polygon", "bbox", "radius"))

galah_group_by(...)

galah_identify(...)

galah_select(..., group = NULL)
}
\description{
These functions are still valid, but have been superseded by more recent
versions. They are designed to be used to supply arguments within
functions from the \link{atlas_} family. Instead you should consider using
piped functions for this same functionality.
}
\details{
Replacements are as follows:

\itemize{
\item \code{\link[=apply_profile]{apply_profile()}} instead of \code{galah_apply_profile()}
\item \code{\link[=filter.data_request]{filter()}} instead of \code{galah_filter()}
\item \code{\link[=geolocate.data_request]{geolocate()}} instead of \code{galah_geolocate()}
\item \code{\link[=group_by.data_request]{select()}} instead of \code{galah_group_by()}
\item \code{\link[=identify.data_request]{select()}} instead of \code{galah_identify()}
\item \code{\link[=select.data_request]{select()}} instead of \code{galah_select()}
}
}
