% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superseded-functions.R
\name{galah_apply_profile}
\alias{galah_apply_profile}
\alias{galah_filter}
\alias{galah_geolocate}
\alias{galah_polygon}
\alias{galah_bbox}
\alias{galah_radius}
\alias{galah_group_by}
\alias{galah_identify}
\alias{galah_select}
\alias{superseded_functions}
\title{Superseded functions}
\usage{
galah_apply_profile(...)

galah_filter(...)

galah_geolocate(..., type = c("polygon", "bbox", "radius"))

galah_polygon(...)

galah_bbox(...)

galah_radius(...)

galah_group_by(...)

galah_identify(...)

galah_select(..., group = NULL)
}
\arguments{
\item{...}{Arguments passed to the function in question, usually (but
not always) using non-standard evaluation.}

\item{type}{\code{string}: one of \code{"polygon"}, \code{"bbox"} or \code{"radius"}. Defaults to
\code{"polygon"}. If \code{type = "polygon"}, a multipolygon will be built via
\code{\link[=geolocate_polygon]{geolocate_polygon()}}. If \code{type = "bbox"}, a multipolygon will be built via
\code{\link[=geolocate_bbox]{geolocate_bbox()}}. The multipolygon is used to narrow a query to the ALA.}

\item{group}{\code{string}: (optional) name of one or more column groups to
include. Valid options are \code{"basic"}, \code{"event"} \code{"taxonomy"}, \code{"media"} and
\code{"assertions"}.}
}
\description{
These functions are still valid, but have been superseded by more recent
versions. They are designed to be used to supply arguments within
functions from the \link{atlas_} family. Instead you should consider using
piped functions for this same functionality.
}
\details{
Replacements are as follows:

\itemize{
\item \code{\link[=apply_profile]{apply_profile()}} instead of \code{galah_apply_profile()}
\item \code{\link[=filter.data_request]{filter()}} instead of \code{galah_filter()}
\item \code{\link[=geolocate]{geolocate()}} instead of \code{galah_geolocate()}
\item \code{\link[=geolocate_polygon]{geolocate_polygon()}} instead of \code{galah_polygon()}
\item \code{\link[=geolocate_bbox]{geolocate_bbox()}} instead of \code{galah_bbox()}
\item \code{\link[=geolocate_radius]{geolocate_radius()}} instead of \code{galah_radius()}
\item \code{\link[=group_by.data_request]{group_by()}} instead of \code{galah_group_by()}
\item \code{\link[=identify.data_request]{identify()}} instead of \code{galah_identify()}
\item \code{\link[=select.data_request]{select()}} instead of \code{galah_select()}
}
}
