% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomic-searches.R
\name{taxonomic_searches}
\alias{taxonomic_searches}
\title{Look up taxon information}
\description{
\code{search_taxa()} allows users to look up taxonomic names before downloading
data from the ALA (using \code{\link[=atlas_occurrences]{atlas_occurrences()}}, \code{\link[=atlas_species]{atlas_species()}} or
\code{\link[=atlas_counts]{atlas_counts()}}). Taxon information returned by \code{search_taxa()} may be
passed to \code{\link[=galah_identify]{galah_identify()}} as the \code{identify} argument of
\code{atlas_} functions. \code{search_taxa()} allows users to disambiguate homonyms
(i.e. where the same name refers to taxa in different clades) prior to
downloading data.

Users can also specify taxonomic levels in a search using a data frame
(tibble). Taxa may be specified using either the \code{specificEpithet} argument
to designate the second element of a Latin binomial,
or the \code{scientificName} argument to specify the
scientific name (which may include the subspecific epithet if required).

\code{search_identifiers()} allows users to look up matching taxonomic names using
their unique \code{taxonConceptID}. In the ALA, all records are associated with
an identifier that uniquely identifies the taxon to which that record belongs.
Once those identifiers
are known, this function allows you to use them to look up further information
on the taxon in question. Effectively this is the inverse function to
\code{\link[=search_taxa]{search_taxa()}}, which takes names and provides identifiers. The resulting
\code{tibble} of taxonomic information can also be passed to \code{\link[=galah_identify]{galah_identify()}} to
filter queries to the specified taxon or taxa.
}
\details{
\code{search_taxa()} returns the taxonomic match of a supplied text string, along
with the following information:
\itemize{
\item \code{search_term}: The search term used by the user. When multiple search
terms are provided in a tibble, these are displayed in this column,
concatenated using \verb{_}.
\item \code{scientific_name}: The taxonomic name matched to the provided search
term, to the lowest identified taxonomic rank.
\item \code{taxon_concept_id}: The unique taxonomic identifier.
\item \code{rank}: The taxonomic rank of the returned result.
\item \code{match_type}: (ALA only) The method of name matching used by the name
matching service. More information can be found on the
\href{https://github.com/AtlasOfLivingAustralia/ala-name-matching?tab=readme-ov-file#understanding-the-name-matching-algorithm}{name matching github repository}.
\item \code{issues}: Any errors returned by the name matching service
(e.g. homonym, indeterminate species match). More information can be found
on the \href{https://github.com/AtlasOfLivingAustralia/ala-name-matching?tab=readme-ov-file#error-types}{name matching github repository}.
\item \verb{taxonomic names} (e.g. \code{kingdom}, \code{phylum}, \code{class}, \code{order},
\code{family}, \code{genus})
}
}
\examples{
# Search using a single string. 
# Note that `search_taxa()` is not case sensitive
search_taxa("Reptilia")

# Search using multiple strings. 
# `search_taxa()` will return one row per taxon
search_taxa("reptilia", "mammalia")

# Search using more detailed strings with authorship information
search_taxa("Acanthocladium F.Muell")

# Specify taxonomic levels in a tibble using "specificEpithet"
search_taxa(tibble::tibble(
  class = "aves", 
  family = "pardalotidae", 
  genus = "pardalotus", 
  specificEpithet = "punctatus"))

# Specify taxonomic levels in a tibble using "scientificName"                    
search_taxa(tibble::tibble(
  family = c("pardalotidae", "maluridae"), 
  scientificName = c("Pardalotus striatus striatus", "malurus cyaneus")))

# Look up a unique taxon identifier
search_identifiers(query = "https://id.biodiversity.org.au/node/apni/2914510")
}
\seealso{
\code{\link[=search_all]{search_all()}} for how to get names if taxonomic identifiers
are already known. \code{\link[=galah_identify]{galah_identify()}}, \code{\link[=galah_select]{galah_select()}}, \code{\link[=galah_filter]{galah_filter()}}, and
\code{\link[=galah_geolocate]{galah_geolocate()}} for ways to restrict the information returned by
\code{\link[=atlas_occurrences]{atlas_occurrences()}} and related functions. \code{\link[=atlas_taxonomy]{atlas_taxonomy()}} to look
up taxonomic trees.
}
