% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyverse.R
\name{tidyverse_functions}
\alias{tidyverse_functions}
\alias{desc}
\alias{unnest}
\title{Non-generic tidyverse functions}
\usage{
desc(...)

unnest(q_obj)
}
\arguments{
\item{...}{column to order by}

\item{q_obj}{An object of class \code{metadata_request}}
}
\value{
description
}
\description{
Several useful functions from \code{{tidyverse}} packages are \code{generic}, meaning
that we can define class-specific versions of those functions and implement
them in galah; examples include \code{filter()}, \code{select()} and \code{group_by()}.
However, there are also functions that are only defined within tidyverse
packages and are not generic. In a few cases we have re-implemented these
functions in \code{galah}. This has the consequence of supporting consistent
syntax with tidyverse, at the cost of potentially introducing conflicts.
This can be avoided by using the \code{::} operator where required (see examples).
}
\details{
The following functions are included:
\itemize{
\item \code{desc()} (\code{dplyr}): use within \code{arrange()} to specify arrangement should be descending
\item \code{unnest()} (\code{tidyr}): use to 'drill down' into nested information on \code{fields}, \code{lists}, \code{profiles}, or \code{taxa}
}

These (\code{galah}) versions all use lazy evaluation.
}
